sdl3_sys::asyncio
Function SDL_GetAsyncIOResult

pub unsafe extern "C" fn SDL_GetAsyncIOResult(
    queue: *mut SDL_AsyncIOQueue,
    outcome: *mut SDL_AsyncIOOutcome,
) -> bool

Query an async I/O task queue for completed tasks.

If a task assigned to this queue has finished, this will return true and fill in outcome with the details of the task. If no task in the queue has finished, this function will return false. This function does not block.

If a task has completed, this function will free its resources and the task pointer will no longer be valid. The task will be removed from the queue.

It is safe for multiple threads to call this function on the same queue at once; a completed task will only go to one of the threads.

Parameters

queue: the async I/O task queue to query.

outcome: details of a finished task will be written here. May not be NULL.

Return value

Returns true if a task has completed, false otherwise.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_WaitAsyncIOResult