sdl3_sys::atomic
Function SDL_CompilerBarrier

pub fn SDL_CompilerBarrier()

Mark a compiler barrier.

A compiler barrier prevents the compiler from reordering reads and writes to globally visible variables across the call.

This macro only prevents the compiler from reordering reads and writes, it does not prevent the CPU from reordering reads and writes. However, all of the atomic operations that modify memory are full memory barriers.

Thread safety

Obviously this macro is safe to use from any thread at any time, but if you find yourself needing this, you are probably dealing with some very sensitive code; be careful!

Availability

This macro is available since SDL 3.2.0.