sdl3_sys::audio
Function SDL_DEFINE_AUDIO_FORMAT

pub const fn SDL_DEFINE_AUDIO_FORMAT(
    signed_: bool,
    bigendian: bool,
    flt: bool,
    size: u8,
) -> SDL_AudioFormat

Define an SDL_AudioFormat value.

SDL does not support custom audio formats, so this macro is not of much use externally, but it can be illustrative as to what the various bits of an SDL_AudioFormat mean.

For example, SDL_AUDIO_S32LE looks like this:

SDL_DEFINE_AUDIO_FORMAT(1, 0, 0, 32)

Parameters

signed: 1 for signed data, 0 for unsigned data.

bigendian: 1 for bigendian data, 0 for littleendian data.

flt: 1 for floating point data, 0 for integer data.

size: number of bits per sample.

Return value

Returns a format value in the style of SDL_AudioFormat.

Thread safety

It is safe to call this macro from any thread.

Availability

This macro is available since SDL 3.2.0.