sdl3_sys::audio
Function SDL_LoadWAV

pub unsafe extern "C" fn SDL_LoadWAV(
    path: *const c_char,
    spec: *mut SDL_AudioSpec,
    audio_buf: *mut *mut Uint8,
    audio_len: *mut Uint32,
) -> bool

Loads a WAV from a file path.

This is a convenience function that is effectively the same as:

SDL_LoadWAV_IO(SDL_IOFromFile(path, "rb"), true, spec, audio_buf, audio_len);

Parameters

path: the file path of the WAV file to open.

spec: a pointer to an SDL_AudioSpec that will be set to the WAVE data’s format details on successful return.

audio_buf: a pointer filled with the audio data, allocated by the function.

audio_len: a pointer filled with the length of the audio data buffer in bytes.

Return value

Returns true on success. audio_buf will be filled with a pointer to an allocated buffer containing the audio data, and audio_len is filled with the length of that audio buffer in bytes.

     This function returns false if the .WAV file cannot be opened,
     uses an unknown data format, or is corrupt; call SDL_GetError()
     for more information.

     When the application is done with the data returned in
     `audio_buf`, it should call SDL_free() to dispose of it.

\threadsafety It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_free

SDL_LoadWAV_IO