sdl3_sys::audio
Function SDL_MixAudio

pub unsafe extern "C" fn SDL_MixAudio(
    dst: *mut Uint8,
    src: *const Uint8,
    format: SDL_AudioFormat,
    len: Uint32,
    volume: c_float,
) -> bool

Mix audio data in a specified format.

This takes an audio buffer src of len bytes of format data and mixes it into dst, performing addition, volume adjustment, and overflow clipping. The buffer pointed to by dst must also be len bytes of format data.

This is provided for convenience – you can mix your own audio data.

Do not use this function for mixing together more than two streams of sample data. The output from repeated application of this function may be distorted by clipping, because there is no accumulator with greater range than the input (not to mention this being an inefficient way of doing it).

It is a common misconception that this function is required to write audio data to an output stream in an audio callback. While you can do that, SDL_MixAudio() is really only needed when you’re mixing a single audio stream with a volume adjustment.

Parameters

dst: the destination for the mixed audio.

src: the source audio buffer to be mixed.

format: the SDL_AudioFormat structure representing the desired audio format.

len: the length of the audio buffer in bytes.

volume: ranges from 0.0 - 1.0, and should be set to 1.0 for full audio volume.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.