sdl3_sys::audio
Function SDL_SetAudioStreamPutCallback

pub unsafe extern "C" fn SDL_SetAudioStreamPutCallback(
    stream: *mut SDL_AudioStream,
    callback: SDL_AudioStreamCallback,
    userdata: *mut c_void,
) -> bool

Set a callback that runs when data is added to an audio stream.

This callback is called after the data is added to the stream, giving the callback the chance to obtain it immediately.

The callback can (optionally) call SDL_GetAudioStreamData() to obtain audio from the stream during this call.

The callback’s approx_request argument is how many bytes of converted data (in the stream’s output format) was provided by the caller, although this may underestimate a little for safety. This value might be less than what is currently available in the stream, if data was already there, and might be less than the caller provided if the stream needs to keep a buffer to aid in resampling. Which means the callback may be provided with zero bytes, and a different amount on each call.

The callback may call SDL_GetAudioStreamAvailable to see the total amount currently available to read from the stream, instead of the total provided by the current call.

The callback is not required to obtain all data. It is allowed to read less or none at all. Anything not read now simply remains in the stream for later access.

Clearing or flushing an audio stream does not call this callback.

This function obtains the stream’s lock, which means any existing callback (get or put) in progress will finish running before setting the new callback.

Setting a NULL function turns off the callback.

Parameters

stream: the audio stream to set the new callback on.

callback: the new callback function to call when data is added to the stream.

userdata: an opaque pointer provided to the callback for its own personal use.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information. This only fails if stream is NULL.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_SetAudioStreamGetCallback