sdl3_sys::camera
Function SDL_GetCameraFormat

pub unsafe extern "C" fn SDL_GetCameraFormat(
    camera: *mut SDL_Camera,
    spec: *mut SDL_CameraSpec,
) -> bool

Get the spec that a camera is using when generating images.

Note that this might not be the native format of the hardware, as SDL might be converting to this format behind the scenes.

If the system is waiting for the user to approve access to the camera, as some platforms require, this will return false, but this isn’t necessarily a fatal error; you should either wait for an SDL_EVENT_CAMERA_DEVICE_APPROVED (or SDL_EVENT_CAMERA_DEVICE_DENIED) event, or poll SDL_GetCameraPermissionState() occasionally until it returns non-zero.

Parameters

camera: opened camera device.

spec: the SDL_CameraSpec to be initialized by this function.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_OpenCamera