sdl3_sys::camera
Function SDL_ReleaseCameraFrame

pub unsafe extern "C" fn SDL_ReleaseCameraFrame(
    camera: *mut SDL_Camera,
    frame: *mut SDL_Surface,
)

Release a frame of video acquired from a camera.

Let the back-end re-use the internal buffer for camera.

This function must be called only on surface objects returned by SDL_AcquireCameraFrame(). This function should be called as quickly as possible after acquisition, as SDL keeps a small FIFO queue of surfaces for video frames; if surfaces aren’t released in a timely manner, SDL may drop upcoming video frames from the camera.

If the app needs to keep the surface for a significant time, they should make a copy of it and release the original.

The app should not use the surface again after calling this function; assume the surface is freed and the pointer is invalid.

Parameters

camera: opened camera device.

frame: the video frame surface to release.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_AcquireCameraFrame