sdl3_sys::clipboard
Function SDL_GetClipboardMimeTypes

pub unsafe extern "C" fn SDL_GetClipboardMimeTypes(
    num_mime_types: *mut usize,
) -> *mut *mut c_char

Retrieve the list of mime types available in the clipboard.

Parameters

num_mime_types: a pointer filled with the number of mime types, may be NULL.

Return value

Returns a null terminated array of strings with mime types, or NULL on failure; call SDL_GetError() for more information. This should be freed with SDL_free() when it is no longer needed.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_SetClipboardData