sdl3_sys::clipboard
Function SDL_SetClipboardData

pub unsafe extern "C" fn SDL_SetClipboardData(
    callback: SDL_ClipboardDataCallback,
    cleanup: SDL_ClipboardCleanupCallback,
    userdata: *mut c_void,
    mime_types: *mut *const c_char,
    num_mime_types: usize,
) -> bool

Offer clipboard data to the OS.

Tell the operating system that the application is offering clipboard data for each of the provided mime-types. Once another application requests the data the callback function will be called, allowing it to generate and respond with the data for the requested mime-type.

The size of text data does not include any terminator, and the text does not need to be null terminated (e.g. you can directly copy a portion of a document).

Parameters

callback: a function pointer to the function that provides the clipboard data.

cleanup: a function pointer to the function that cleans up the clipboard data.

userdata: an opaque pointer that will be forwarded to the callbacks.

mime_types: a list of mime-types that are being offered.

num_mime_types: the number of mime-types in the mime_types list.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_ClearClipboardData

SDL_GetClipboardData

SDL_HasClipboardData