sdl3_sys::error
Function SDL_SetError

pub unsafe extern "C" fn SDL_SetError(fmt: *const c_char, ...) -> bool

Set the SDL error message for the current thread.

Calling this function will replace any previous error message that was set.

This function always returns false, since SDL frequently uses false to signify a failing result, leading to this idiom:

if (error_code) {
    return SDL_SetError("This operation has failed: %d", error_code);
}

Parameters

fmt: a printf()-style message format string.

...: additional parameters matching % tokens in the fmt string, if any.

Return value

Returns false.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_ClearError

SDL_GetError

SDL_SetErrorV