sdl3_sys::events
Struct SDL_EventType

#[repr(transparent)]
pub struct SDL_EventType(pub Uint32);

The types of events that can be delivered.

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

FIRST	SDL_EVENT_FIRST	Unused (do not remove)

QUIT	SDL_EVENT_QUIT	User-requested quit

TERMINATING	SDL_EVENT_TERMINATING	The application is being terminated by the OS. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillTerminate() Called on Android in onDestroy()

LOW_MEMORY	SDL_EVENT_LOW_MEMORY	The application is low on memory, free memory if possible. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidReceiveMemoryWarning() Called on Android in onTrimMemory()

WILL_ENTER_BACKGROUND	SDL_EVENT_WILL_ENTER_BACKGROUND	The application is about to enter the background. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillResignActive() Called on Android in onPause()

DID_ENTER_BACKGROUND	SDL_EVENT_DID_ENTER_BACKGROUND	The application did enter the background and may not get CPU for some time. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidEnterBackground() Called on Android in onPause()

WILL_ENTER_FOREGROUND	SDL_EVENT_WILL_ENTER_FOREGROUND	The application is about to enter the foreground. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillEnterForeground() Called on Android in onResume()

DID_ENTER_FOREGROUND	SDL_EVENT_DID_ENTER_FOREGROUND	The application is now interactive. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidBecomeActive() Called on Android in onResume()

LOCALE_CHANGED	SDL_EVENT_LOCALE_CHANGED	The user’s locale preferences have changed.

SYSTEM_THEME_CHANGED	SDL_EVENT_SYSTEM_THEME_CHANGED	The system theme changed

DISPLAY_ORIENTATION	SDL_EVENT_DISPLAY_ORIENTATION	Display orientation has changed to data1

DISPLAY_ADDED	SDL_EVENT_DISPLAY_ADDED	Display has been added to the system

DISPLAY_REMOVED	SDL_EVENT_DISPLAY_REMOVED	Display has been removed from the system

DISPLAY_MOVED	SDL_EVENT_DISPLAY_MOVED	Display has changed position

DISPLAY_DESKTOP_MODE_CHANGED	SDL_EVENT_DISPLAY_DESKTOP_MODE_CHANGED	Display has changed desktop mode

DISPLAY_CURRENT_MODE_CHANGED	SDL_EVENT_DISPLAY_CURRENT_MODE_CHANGED	Display has changed current mode

DISPLAY_CONTENT_SCALE_CHANGED	SDL_EVENT_DISPLAY_CONTENT_SCALE_CHANGED	Display has changed content scale

DISPLAY_FIRST	SDL_EVENT_DISPLAY_FIRST

DISPLAY_LAST	SDL_EVENT_DISPLAY_LAST

WINDOW_SHOWN	SDL_EVENT_WINDOW_SHOWN	Window has been shown

WINDOW_HIDDEN	SDL_EVENT_WINDOW_HIDDEN	Window has been hidden

WINDOW_EXPOSED	SDL_EVENT_WINDOW_EXPOSED	Window has been exposed and should be redrawn, and can be redrawn directly from event watchers for this event

WINDOW_MOVED	SDL_EVENT_WINDOW_MOVED	Window has been moved to data1, data2

WINDOW_RESIZED	SDL_EVENT_WINDOW_RESIZED	Window has been resized to data1xdata2

WINDOW_PIXEL_SIZE_CHANGED	SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED	The pixel size of the window has changed to data1xdata2

WINDOW_METAL_VIEW_RESIZED	SDL_EVENT_WINDOW_METAL_VIEW_RESIZED	The pixel size of a Metal view associated with the window has changed

WINDOW_MINIMIZED	SDL_EVENT_WINDOW_MINIMIZED	Window has been minimized

WINDOW_MAXIMIZED	SDL_EVENT_WINDOW_MAXIMIZED	Window has been maximized

WINDOW_RESTORED	SDL_EVENT_WINDOW_RESTORED	Window has been restored to normal size and position

WINDOW_MOUSE_ENTER	SDL_EVENT_WINDOW_MOUSE_ENTER	Window has gained mouse focus

WINDOW_MOUSE_LEAVE	SDL_EVENT_WINDOW_MOUSE_LEAVE	Window has lost mouse focus

WINDOW_FOCUS_GAINED	SDL_EVENT_WINDOW_FOCUS_GAINED	Window has gained keyboard focus

WINDOW_FOCUS_LOST	SDL_EVENT_WINDOW_FOCUS_LOST	Window has lost keyboard focus

WINDOW_CLOSE_REQUESTED	SDL_EVENT_WINDOW_CLOSE_REQUESTED	The window manager requests that the window be closed

WINDOW_HIT_TEST	SDL_EVENT_WINDOW_HIT_TEST	Window had a hit test that wasn’t SDL_HITTEST_NORMAL

WINDOW_ICCPROF_CHANGED	SDL_EVENT_WINDOW_ICCPROF_CHANGED	The ICC profile of the window’s display has changed

WINDOW_DISPLAY_CHANGED	SDL_EVENT_WINDOW_DISPLAY_CHANGED	Window has been moved to display data1

WINDOW_DISPLAY_SCALE_CHANGED	SDL_EVENT_WINDOW_DISPLAY_SCALE_CHANGED	Window display scale has been changed

WINDOW_SAFE_AREA_CHANGED	SDL_EVENT_WINDOW_SAFE_AREA_CHANGED	The window safe area has been changed

WINDOW_OCCLUDED	SDL_EVENT_WINDOW_OCCLUDED	The window has been occluded

WINDOW_ENTER_FULLSCREEN	SDL_EVENT_WINDOW_ENTER_FULLSCREEN	The window has entered fullscreen mode

WINDOW_LEAVE_FULLSCREEN	SDL_EVENT_WINDOW_LEAVE_FULLSCREEN	The window has left fullscreen mode

WINDOW_DESTROYED	SDL_EVENT_WINDOW_DESTROYED	The window with the associated ID is being or has been destroyed. If this message is being handled in an event watcher, the window handle is still valid and can still be used to retrieve any properties associated with the window. Otherwise, the handle has already been destroyed and all resources associated with it are invalid

WINDOW_HDR_STATE_CHANGED	SDL_EVENT_WINDOW_HDR_STATE_CHANGED	Window HDR properties have changed

WINDOW_FIRST	SDL_EVENT_WINDOW_FIRST

WINDOW_LAST	SDL_EVENT_WINDOW_LAST

KEY_DOWN	SDL_EVENT_KEY_DOWN	Key pressed

KEY_UP	SDL_EVENT_KEY_UP	Key released

TEXT_EDITING	SDL_EVENT_TEXT_EDITING	Keyboard text editing (composition)

TEXT_INPUT	SDL_EVENT_TEXT_INPUT	Keyboard text input

KEYMAP_CHANGED	SDL_EVENT_KEYMAP_CHANGED	Keymap changed due to a system event such as an input language or keyboard layout change.

KEYBOARD_ADDED	SDL_EVENT_KEYBOARD_ADDED	A new keyboard has been inserted into the system

KEYBOARD_REMOVED	SDL_EVENT_KEYBOARD_REMOVED	A keyboard has been removed

TEXT_EDITING_CANDIDATES	SDL_EVENT_TEXT_EDITING_CANDIDATES	Keyboard text editing candidates

MOUSE_MOTION	SDL_EVENT_MOUSE_MOTION	Mouse moved

MOUSE_BUTTON_DOWN	SDL_EVENT_MOUSE_BUTTON_DOWN	Mouse button pressed

MOUSE_BUTTON_UP	SDL_EVENT_MOUSE_BUTTON_UP	Mouse button released

MOUSE_WHEEL	SDL_EVENT_MOUSE_WHEEL	Mouse wheel motion

MOUSE_ADDED	SDL_EVENT_MOUSE_ADDED	A new mouse has been inserted into the system

MOUSE_REMOVED	SDL_EVENT_MOUSE_REMOVED	A mouse has been removed

JOYSTICK_AXIS_MOTION	SDL_EVENT_JOYSTICK_AXIS_MOTION	Joystick axis motion

JOYSTICK_BALL_MOTION	SDL_EVENT_JOYSTICK_BALL_MOTION	Joystick trackball motion

JOYSTICK_HAT_MOTION	SDL_EVENT_JOYSTICK_HAT_MOTION	Joystick hat position change

JOYSTICK_BUTTON_DOWN	SDL_EVENT_JOYSTICK_BUTTON_DOWN	Joystick button pressed

JOYSTICK_BUTTON_UP	SDL_EVENT_JOYSTICK_BUTTON_UP	Joystick button released

JOYSTICK_ADDED	SDL_EVENT_JOYSTICK_ADDED	A new joystick has been inserted into the system

JOYSTICK_REMOVED	SDL_EVENT_JOYSTICK_REMOVED	An opened joystick has been removed

JOYSTICK_BATTERY_UPDATED	SDL_EVENT_JOYSTICK_BATTERY_UPDATED	Joystick battery level change

JOYSTICK_UPDATE_COMPLETE	SDL_EVENT_JOYSTICK_UPDATE_COMPLETE	Joystick update is complete

GAMEPAD_AXIS_MOTION	SDL_EVENT_GAMEPAD_AXIS_MOTION	Gamepad axis motion

GAMEPAD_BUTTON_DOWN	SDL_EVENT_GAMEPAD_BUTTON_DOWN	Gamepad button pressed

GAMEPAD_BUTTON_UP	SDL_EVENT_GAMEPAD_BUTTON_UP	Gamepad button released

GAMEPAD_ADDED	SDL_EVENT_GAMEPAD_ADDED	A new gamepad has been inserted into the system

GAMEPAD_REMOVED	SDL_EVENT_GAMEPAD_REMOVED	A gamepad has been removed

GAMEPAD_REMAPPED	SDL_EVENT_GAMEPAD_REMAPPED	The gamepad mapping was updated

GAMEPAD_TOUCHPAD_DOWN	SDL_EVENT_GAMEPAD_TOUCHPAD_DOWN	Gamepad touchpad was touched

GAMEPAD_TOUCHPAD_MOTION	SDL_EVENT_GAMEPAD_TOUCHPAD_MOTION	Gamepad touchpad finger was moved

GAMEPAD_TOUCHPAD_UP	SDL_EVENT_GAMEPAD_TOUCHPAD_UP	Gamepad touchpad finger was lifted

GAMEPAD_SENSOR_UPDATE	SDL_EVENT_GAMEPAD_SENSOR_UPDATE	Gamepad sensor was updated

GAMEPAD_UPDATE_COMPLETE	SDL_EVENT_GAMEPAD_UPDATE_COMPLETE	Gamepad update is complete

GAMEPAD_STEAM_HANDLE_UPDATED	SDL_EVENT_GAMEPAD_STEAM_HANDLE_UPDATED	Gamepad Steam handle has changed

FINGER_DOWN	SDL_EVENT_FINGER_DOWN

FINGER_UP	SDL_EVENT_FINGER_UP

FINGER_MOTION	SDL_EVENT_FINGER_MOTION

FINGER_CANCELED	SDL_EVENT_FINGER_CANCELED

CLIPBOARD_UPDATE	SDL_EVENT_CLIPBOARD_UPDATE	The clipboard or primary selection changed

DROP_FILE	SDL_EVENT_DROP_FILE	The system requests a file open

DROP_TEXT	SDL_EVENT_DROP_TEXT	text/plain drag-and-drop event

DROP_BEGIN	SDL_EVENT_DROP_BEGIN	A new set of drops is beginning (NULL filename)

DROP_COMPLETE	SDL_EVENT_DROP_COMPLETE	Current set of drops is now complete (NULL filename)

DROP_POSITION	SDL_EVENT_DROP_POSITION	Position while moving over the window

AUDIO_DEVICE_ADDED	SDL_EVENT_AUDIO_DEVICE_ADDED	A new audio device is available

AUDIO_DEVICE_REMOVED	SDL_EVENT_AUDIO_DEVICE_REMOVED	An audio device has been removed.

AUDIO_DEVICE_FORMAT_CHANGED	SDL_EVENT_AUDIO_DEVICE_FORMAT_CHANGED	An audio device’s format has been changed by the system.

SENSOR_UPDATE	SDL_EVENT_SENSOR_UPDATE	A sensor was updated

PEN_PROXIMITY_IN	SDL_EVENT_PEN_PROXIMITY_IN	Pressure-sensitive pen has become available

PEN_PROXIMITY_OUT	SDL_EVENT_PEN_PROXIMITY_OUT	Pressure-sensitive pen has become unavailable

PEN_DOWN	SDL_EVENT_PEN_DOWN	Pressure-sensitive pen touched drawing surface

PEN_UP	SDL_EVENT_PEN_UP	Pressure-sensitive pen stopped touching drawing surface

PEN_BUTTON_DOWN	SDL_EVENT_PEN_BUTTON_DOWN	Pressure-sensitive pen button pressed

PEN_BUTTON_UP	SDL_EVENT_PEN_BUTTON_UP	Pressure-sensitive pen button released

PEN_MOTION	SDL_EVENT_PEN_MOTION	Pressure-sensitive pen is moving on the tablet

PEN_AXIS	SDL_EVENT_PEN_AXIS	Pressure-sensitive pen angle/pressure/etc changed

CAMERA_DEVICE_ADDED	SDL_EVENT_CAMERA_DEVICE_ADDED	A new camera device is available

CAMERA_DEVICE_REMOVED	SDL_EVENT_CAMERA_DEVICE_REMOVED	A camera device has been removed.

CAMERA_DEVICE_APPROVED	SDL_EVENT_CAMERA_DEVICE_APPROVED	A camera device has been approved for use by the user.

CAMERA_DEVICE_DENIED	SDL_EVENT_CAMERA_DEVICE_DENIED	A camera device has been denied for use by the user.

RENDER_TARGETS_RESET	SDL_EVENT_RENDER_TARGETS_RESET	The render targets have been reset and their contents need to be updated

RENDER_DEVICE_RESET	SDL_EVENT_RENDER_DEVICE_RESET	The device has been reset and all textures need to be recreated

RENDER_DEVICE_LOST	SDL_EVENT_RENDER_DEVICE_LOST	The device has been lost and can’t be recovered.

PRIVATE0	SDL_EVENT_PRIVATE0

PRIVATE1	SDL_EVENT_PRIVATE1

PRIVATE2	SDL_EVENT_PRIVATE2

PRIVATE3	SDL_EVENT_PRIVATE3

POLL_SENTINEL	SDL_EVENT_POLL_SENTINEL	Signals the end of an event poll cycle

USER	SDL_EVENT_USER	Events SDL_EVENT_USER through SDL_EVENT_LAST are for your use, and should be allocated with SDL_RegisterEvents()

LAST	SDL_EVENT_LAST	* This last event is only for bounding internal arrays

ENUM_PADDING	SDL_EVENT_ENUM_PADDING

Tuple Fields
0: Uint32
Implementations

impl SDL_EventType

pub const FIRST: Self

Unused (do not remove)

pub const QUIT: Self

User-requested quit

pub const TERMINATING: Self

The application is being terminated by the OS. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillTerminate() Called on Android in onDestroy()

pub const LOW_MEMORY: Self

The application is low on memory, free memory if possible. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidReceiveMemoryWarning() Called on Android in onTrimMemory()

pub const WILL_ENTER_BACKGROUND: Self

The application is about to enter the background. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillResignActive() Called on Android in onPause()

pub const DID_ENTER_BACKGROUND: Self

The application did enter the background and may not get CPU for some time. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidEnterBackground() Called on Android in onPause()

pub const WILL_ENTER_FOREGROUND: Self

The application is about to enter the foreground. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationWillEnterForeground() Called on Android in onResume()

pub const DID_ENTER_FOREGROUND: Self

The application is now interactive. This event must be handled in a callback set with SDL_AddEventWatch(). Called on iOS in applicationDidBecomeActive() Called on Android in onResume()

pub const LOCALE_CHANGED: Self

The user’s locale preferences have changed.

pub const SYSTEM_THEME_CHANGED: Self

The system theme changed

pub const DISPLAY_ORIENTATION: Self

Display orientation has changed to data1

pub const DISPLAY_ADDED: Self

Display has been added to the system

pub const DISPLAY_REMOVED: Self

Display has been removed from the system

pub const DISPLAY_MOVED: Self

Display has changed position

pub const DISPLAY_DESKTOP_MODE_CHANGED: Self

Display has changed desktop mode

pub const DISPLAY_CURRENT_MODE_CHANGED: Self

Display has changed current mode

pub const DISPLAY_CONTENT_SCALE_CHANGED: Self

Display has changed content scale

pub const DISPLAY_FIRST: Self = SDL_EVENT_DISPLAY_ORIENTATION

pub const DISPLAY_LAST: Self = SDL_EVENT_DISPLAY_CONTENT_SCALE_CHANGED

pub const WINDOW_SHOWN: Self

Window has been shown

pub const WINDOW_HIDDEN: Self

Window has been hidden

pub const WINDOW_EXPOSED: Self

Window has been exposed and should be redrawn, and can be redrawn directly from event watchers for this event

pub const WINDOW_MOVED: Self

Window has been moved to data1, data2

pub const WINDOW_RESIZED: Self

Window has been resized to data1xdata2

pub const WINDOW_PIXEL_SIZE_CHANGED: Self

The pixel size of the window has changed to data1xdata2

pub const WINDOW_METAL_VIEW_RESIZED: Self

The pixel size of a Metal view associated with the window has changed

pub const WINDOW_MINIMIZED: Self

Window has been minimized

pub const WINDOW_MAXIMIZED: Self

Window has been maximized

pub const WINDOW_RESTORED: Self

Window has been restored to normal size and position

pub const WINDOW_MOUSE_ENTER: Self

Window has gained mouse focus

pub const WINDOW_MOUSE_LEAVE: Self

Window has lost mouse focus

pub const WINDOW_FOCUS_GAINED: Self

Window has gained keyboard focus

pub const WINDOW_FOCUS_LOST: Self

Window has lost keyboard focus

pub const WINDOW_CLOSE_REQUESTED: Self

The window manager requests that the window be closed

pub const WINDOW_HIT_TEST: Self

Window had a hit test that wasn’t SDL_HITTEST_NORMAL

pub const WINDOW_ICCPROF_CHANGED: Self

The ICC profile of the window’s display has changed

pub const WINDOW_DISPLAY_CHANGED: Self

Window has been moved to display data1

pub const WINDOW_DISPLAY_SCALE_CHANGED: Self

Window display scale has been changed

pub const WINDOW_SAFE_AREA_CHANGED: Self

The window safe area has been changed

pub const WINDOW_OCCLUDED: Self

The window has been occluded

pub const WINDOW_ENTER_FULLSCREEN: Self

The window has entered fullscreen mode

pub const WINDOW_LEAVE_FULLSCREEN: Self

The window has left fullscreen mode

pub const WINDOW_DESTROYED: Self

The window with the associated ID is being or has been destroyed. If this message is being handled in an event watcher, the window handle is still valid and can still be used to retrieve any properties associated with the window. Otherwise, the handle has already been destroyed and all resources associated with it are invalid

pub const WINDOW_HDR_STATE_CHANGED: Self

Window HDR properties have changed

pub const WINDOW_FIRST: Self = SDL_EVENT_WINDOW_SHOWN

pub const WINDOW_LAST: Self = SDL_EVENT_WINDOW_HDR_STATE_CHANGED

pub const KEY_DOWN: Self

Key pressed

pub const KEY_UP: Self

Key released

pub const TEXT_EDITING: Self

Keyboard text editing (composition)

pub const TEXT_INPUT: Self

Keyboard text input

pub const KEYMAP_CHANGED: Self

Keymap changed due to a system event such as an input language or keyboard layout change.

pub const KEYBOARD_ADDED: Self

A new keyboard has been inserted into the system

pub const KEYBOARD_REMOVED: Self

A keyboard has been removed

pub const TEXT_EDITING_CANDIDATES: Self

Keyboard text editing candidates

pub const MOUSE_MOTION: Self

Mouse moved

pub const MOUSE_BUTTON_DOWN: Self

Mouse button pressed

pub const MOUSE_BUTTON_UP: Self

Mouse button released

pub const MOUSE_WHEEL: Self

Mouse wheel motion

pub const MOUSE_ADDED: Self

A new mouse has been inserted into the system

pub const MOUSE_REMOVED: Self

A mouse has been removed

pub const JOYSTICK_AXIS_MOTION: Self

Joystick axis motion

pub const JOYSTICK_BALL_MOTION: Self

Joystick trackball motion

pub const JOYSTICK_HAT_MOTION: Self

Joystick hat position change

pub const JOYSTICK_BUTTON_DOWN: Self

Joystick button pressed

pub const JOYSTICK_BUTTON_UP: Self

Joystick button released

pub const JOYSTICK_ADDED: Self

A new joystick has been inserted into the system

pub const JOYSTICK_REMOVED: Self

An opened joystick has been removed

pub const JOYSTICK_BATTERY_UPDATED: Self

Joystick battery level change

pub const JOYSTICK_UPDATE_COMPLETE: Self

Joystick update is complete

pub const GAMEPAD_AXIS_MOTION: Self

Gamepad axis motion

pub const GAMEPAD_BUTTON_DOWN: Self

Gamepad button pressed

pub const GAMEPAD_BUTTON_UP: Self

Gamepad button released

pub const GAMEPAD_ADDED: Self

A new gamepad has been inserted into the system

pub const GAMEPAD_REMOVED: Self

A gamepad has been removed

pub const GAMEPAD_REMAPPED: Self

The gamepad mapping was updated

pub const GAMEPAD_TOUCHPAD_DOWN: Self

Gamepad touchpad was touched

pub const GAMEPAD_TOUCHPAD_MOTION: Self

Gamepad touchpad finger was moved

pub const GAMEPAD_TOUCHPAD_UP: Self

Gamepad touchpad finger was lifted

pub const GAMEPAD_SENSOR_UPDATE: Self

Gamepad sensor was updated

pub const GAMEPAD_UPDATE_COMPLETE: Self

Gamepad update is complete

pub const GAMEPAD_STEAM_HANDLE_UPDATED: Self

Gamepad Steam handle has changed

pub const FINGER_DOWN: Self

pub const FINGER_UP: Self

pub const FINGER_MOTION: Self

pub const FINGER_CANCELED: Self

pub const CLIPBOARD_UPDATE: Self

The clipboard or primary selection changed

pub const DROP_FILE: Self

The system requests a file open

pub const DROP_TEXT: Self

text/plain drag-and-drop event

pub const DROP_BEGIN: Self

A new set of drops is beginning (NULL filename)

pub const DROP_COMPLETE: Self

Current set of drops is now complete (NULL filename)

pub const DROP_POSITION: Self

Position while moving over the window

pub const AUDIO_DEVICE_ADDED: Self

A new audio device is available

pub const AUDIO_DEVICE_REMOVED: Self

An audio device has been removed.

pub const AUDIO_DEVICE_FORMAT_CHANGED: Self

An audio device’s format has been changed by the system.

pub const SENSOR_UPDATE: Self

A sensor was updated

pub const PEN_PROXIMITY_IN: Self

Pressure-sensitive pen has become available

pub const PEN_PROXIMITY_OUT: Self

Pressure-sensitive pen has become unavailable

pub const PEN_DOWN: Self

Pressure-sensitive pen touched drawing surface

pub const PEN_UP: Self

Pressure-sensitive pen stopped touching drawing surface

pub const PEN_BUTTON_DOWN: Self

Pressure-sensitive pen button pressed

pub const PEN_BUTTON_UP: Self

Pressure-sensitive pen button released

pub const PEN_MOTION: Self

Pressure-sensitive pen is moving on the tablet

pub const PEN_AXIS: Self

Pressure-sensitive pen angle/pressure/etc changed

pub const CAMERA_DEVICE_ADDED: Self

A new camera device is available

pub const CAMERA_DEVICE_REMOVED: Self

A camera device has been removed.

pub const CAMERA_DEVICE_APPROVED: Self

A camera device has been approved for use by the user.

pub const CAMERA_DEVICE_DENIED: Self

A camera device has been denied for use by the user.

pub const RENDER_TARGETS_RESET: Self

The render targets have been reset and their contents need to be updated

pub const RENDER_DEVICE_RESET: Self

The device has been reset and all textures need to be recreated

pub const RENDER_DEVICE_LOST: Self

The device has been lost and can’t be recovered.

pub const PRIVATE0: Self

pub const PRIVATE1: Self

pub const PRIVATE2: Self

pub const PRIVATE3: Self

pub const POLL_SENTINEL: Self

Signals the end of an event poll cycle

pub const USER: Self

Events SDL_EVENT_USER through SDL_EVENT_LAST are for your use, and should be allocated with SDL_RegisterEvents()

pub const LAST: Self

This last event is only for bounding internal arrays

pub const ENUM_PADDING: Self