sdl3_sys::events
Union SDL_Event

#[repr(C)]
pub union SDL_Event {

    pub type: Uint32,
    pub common: SDL_CommonEvent,
    pub display: SDL_DisplayEvent,
    pub window: SDL_WindowEvent,
    pub kdevice: SDL_KeyboardDeviceEvent,
    pub key: SDL_KeyboardEvent,
    pub edit: SDL_TextEditingEvent,
    pub edit_candidates: SDL_TextEditingCandidatesEvent,
    pub text: SDL_TextInputEvent,
    pub mdevice: SDL_MouseDeviceEvent,
    pub motion: SDL_MouseMotionEvent,
    pub button: SDL_MouseButtonEvent,
    pub wheel: SDL_MouseWheelEvent,
    pub jdevice: SDL_JoyDeviceEvent,
    pub jaxis: SDL_JoyAxisEvent,
    pub jball: SDL_JoyBallEvent,
    pub jhat: SDL_JoyHatEvent,
    pub jbutton: SDL_JoyButtonEvent,
    pub jbattery: SDL_JoyBatteryEvent,
    pub gdevice: SDL_GamepadDeviceEvent,
    pub gaxis: SDL_GamepadAxisEvent,
    pub gbutton: SDL_GamepadButtonEvent,
    pub gtouchpad: SDL_GamepadTouchpadEvent,
    pub gsensor: SDL_GamepadSensorEvent,
    pub adevice: SDL_AudioDeviceEvent,
    pub cdevice: SDL_CameraDeviceEvent,
    pub sensor: SDL_SensorEvent,
    pub quit: SDL_QuitEvent,
    pub user: SDL_UserEvent,
    pub tfinger: SDL_TouchFingerEvent,
    pub pproximity: SDL_PenProximityEvent,
    pub ptouch: SDL_PenTouchEvent,
    pub pmotion: SDL_PenMotionEvent,
    pub pbutton: SDL_PenButtonEvent,
    pub paxis: SDL_PenAxisEvent,
    pub render: SDL_RenderEvent,
    pub drop: SDL_DropEvent,
    pub clipboard: SDL_ClipboardEvent,
    pub padding: [Uint8; 128],

}

The structure for all events in SDL.

The SDL_Event structure is the core of all event handling in SDL. SDL_Event is a union of all event structures used in SDL.

Availability

This struct is available since SDL 3.2.0.

Fields
type: Uint32

Event type, shared with all events, Uint32 to cover user events which are not in the SDL_EventType enumeration

common: SDL_CommonEvent

Common event data

display: SDL_DisplayEvent

Display event data

window: SDL_WindowEvent

Window event data

kdevice: SDL_KeyboardDeviceEvent

Keyboard device change event data

key: SDL_KeyboardEvent

Keyboard event data

edit: SDL_TextEditingEvent

Text editing event data

edit_candidates: SDL_TextEditingCandidatesEvent

Text editing candidates event data

text: SDL_TextInputEvent

Text input event data

mdevice: SDL_MouseDeviceEvent

Mouse device change event data

motion: SDL_MouseMotionEvent

Mouse motion event data

button: SDL_MouseButtonEvent

Mouse button event data

wheel: SDL_MouseWheelEvent

Mouse wheel event data

jdevice: SDL_JoyDeviceEvent

Joystick device change event data

jaxis: SDL_JoyAxisEvent

Joystick axis event data

jball: SDL_JoyBallEvent

Joystick ball event data

jhat: SDL_JoyHatEvent

Joystick hat event data

jbutton: SDL_JoyButtonEvent

Joystick button event data

jbattery: SDL_JoyBatteryEvent

Joystick battery event data

gdevice: SDL_GamepadDeviceEvent

Gamepad device event data

gaxis: SDL_GamepadAxisEvent

Gamepad axis event data

gbutton: SDL_GamepadButtonEvent

Gamepad button event data

gtouchpad: SDL_GamepadTouchpadEvent

Gamepad touchpad event data

gsensor: SDL_GamepadSensorEvent

Gamepad sensor event data

adevice: SDL_AudioDeviceEvent

Audio device event data

cdevice: SDL_CameraDeviceEvent

Camera device event data

sensor: SDL_SensorEvent

Sensor event data

quit: SDL_QuitEvent

Quit request event data

user: SDL_UserEvent

Custom event data

tfinger: SDL_TouchFingerEvent

Touch finger event data

pproximity: SDL_PenProximityEvent

Pen proximity event data

ptouch: SDL_PenTouchEvent

Pen tip touching event data

pmotion: SDL_PenMotionEvent

Pen motion event data

pbutton: SDL_PenButtonEvent

Pen button event data

paxis: SDL_PenAxisEvent

Pen axis event data

render: SDL_RenderEvent

Render event data

drop: SDL_DropEvent

Drag and drop event data

clipboard: SDL_ClipboardEvent

Clipboard event data

padding: [Uint8; 128]