sdl3_sys
Module ffi

Extra ffi types for sdl3-sys

Enums

VaList

Equivalent to C’s va_list type. Enable the nightly feature and compile with the nightly compiler to make this an alias of core::ffi::VaList. Otherwise, this type can’t be instantiated.

Type Aliases

c_wchar_t

Equivalent to C’s wchar_t type. This is u16 on Windows and u32 otherwise. Enable a use-libc-* feature to make this an alias of libc::wchar_t.