sdl3_sys::filesystem
Type Alias SDL_EnumerateDirectoryCallback

pub type SDL_EnumerateDirectoryCallback = Option<unsafe extern "C" fn(userdata: *mut c_void, dirname: *const c_char, fname: *const c_char) -> SDL_EnumerationResult>;

Callback for directory enumeration.

Enumeration of directory entries will continue until either all entries have been provided to the callback, or the callback has requested a stop through its return value.

Returning SDL_ENUM_CONTINUE will let enumeration proceed, calling the callback with further entries. SDL_ENUM_SUCCESS and SDL_ENUM_FAILURE will terminate the enumeration early, and dictate the return value of the enumeration function itself.

dirname is guaranteed to end with a path separator (‘\’ on Windows, ‘/’ on most other platforms).

Parameters

userdata: an app-controlled pointer that is passed to the callback.

dirname: the directory that is being enumerated.

fname: the next entry in the enumeration.

Return value

Returns how the enumeration should proceed.

Availability

This datatype is available since SDL 3.2.0.

See also

SDL_EnumerateDirectory

Aliased Type

enum SDL_EnumerateDirectoryCallback {
    None,
    Some(unsafe extern "C" fn(_: *mut c_void, _: *const i8, _: *const i8) -> SDL_EnumerationResult),
}

Variants

None

No value.

Some(unsafe extern "C" fn(_: *mut c_void, _: *const i8, _: *const i8) -> SDL_EnumerationResult)

Some value of type T.