sdl3_sys::gamepad
Function SDL_GetGamepadSteamHandle

pub unsafe extern "C" fn SDL_GetGamepadSteamHandle(
    gamepad: *mut SDL_Gamepad,
) -> Uint64

Get the Steam Input handle of an opened gamepad, if available.

Returns an InputHandle_t for the gamepad that can be used with Steam Input API: https://partner.steamgames.com/doc/api/ISteamInput

Parameters

gamepad: the gamepad object to query.

Return value

Returns the gamepad handle, or 0 if unavailable.

Availability

This function is available since SDL 3.2.0.