sdl3_sys::gamepad
Struct SDL_GamepadButton

#[repr(transparent)]
pub struct SDL_GamepadButton(pub c_int);

The list of buttons available on a gamepad

For controllers that use a diamond pattern for the face buttons, the south/east/west/north buttons below correspond to the locations in the diamond pattern. For Xbox controllers, this would be A/B/X/Y, for Nintendo Switch controllers, this would be B/A/Y/X, for PlayStation controllers this would be Cross/Circle/Square/Triangle.

For controllers that don’t use a diamond pattern for the face buttons, the south/east/west/north buttons indicate the buttons labeled A, B, C, D, or 1, 2, 3, 4, or for controllers that aren’t labeled, they are the primary, secondary, etc. buttons.

The activate action is often the south button and the cancel action is often the east button, but in some regions this is reversed, so your game should allow remapping actions based on user preferences.

You can query the labels for the face buttons using SDL_GetGamepadButtonLabel()

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

INVALID	SDL_GAMEPAD_BUTTON_INVALID

SOUTH	SDL_GAMEPAD_BUTTON_SOUTH	Bottom face button (e.g. Xbox A button)

EAST	SDL_GAMEPAD_BUTTON_EAST	Right face button (e.g. Xbox B button)

WEST	SDL_GAMEPAD_BUTTON_WEST	Left face button (e.g. Xbox X button)

NORTH	SDL_GAMEPAD_BUTTON_NORTH	Top face button (e.g. Xbox Y button)

BACK	SDL_GAMEPAD_BUTTON_BACK

GUIDE	SDL_GAMEPAD_BUTTON_GUIDE

START	SDL_GAMEPAD_BUTTON_START

LEFT_STICK	SDL_GAMEPAD_BUTTON_LEFT_STICK

RIGHT_STICK	SDL_GAMEPAD_BUTTON_RIGHT_STICK

LEFT_SHOULDER	SDL_GAMEPAD_BUTTON_LEFT_SHOULDER

RIGHT_SHOULDER	SDL_GAMEPAD_BUTTON_RIGHT_SHOULDER

DPAD_UP	SDL_GAMEPAD_BUTTON_DPAD_UP

DPAD_DOWN	SDL_GAMEPAD_BUTTON_DPAD_DOWN

DPAD_LEFT	SDL_GAMEPAD_BUTTON_DPAD_LEFT

DPAD_RIGHT	SDL_GAMEPAD_BUTTON_DPAD_RIGHT

MISC1	SDL_GAMEPAD_BUTTON_MISC1	Additional button (e.g. Xbox Series X share button, PS5 microphone button, Nintendo Switch Pro capture button, Amazon Luna microphone button, Google Stadia capture button)

RIGHT_PADDLE1	SDL_GAMEPAD_BUTTON_RIGHT_PADDLE1	Upper or primary paddle, under your right hand (e.g. Xbox Elite paddle P1)

LEFT_PADDLE1	SDL_GAMEPAD_BUTTON_LEFT_PADDLE1	Upper or primary paddle, under your left hand (e.g. Xbox Elite paddle P3)

RIGHT_PADDLE2	SDL_GAMEPAD_BUTTON_RIGHT_PADDLE2	Lower or secondary paddle, under your right hand (e.g. Xbox Elite paddle P2)

LEFT_PADDLE2	SDL_GAMEPAD_BUTTON_LEFT_PADDLE2	Lower or secondary paddle, under your left hand (e.g. Xbox Elite paddle P4)

TOUCHPAD	SDL_GAMEPAD_BUTTON_TOUCHPAD	PS4/PS5 touchpad button

MISC2	SDL_GAMEPAD_BUTTON_MISC2	Additional button

MISC3	SDL_GAMEPAD_BUTTON_MISC3	Additional button

MISC4	SDL_GAMEPAD_BUTTON_MISC4	Additional button

MISC5	SDL_GAMEPAD_BUTTON_MISC5	Additional button

MISC6	SDL_GAMEPAD_BUTTON_MISC6	Additional button

COUNT	SDL_GAMEPAD_BUTTON_COUNT

Tuple Fields
0: c_int
Implementations

impl SDL_GamepadButton

pub const INVALID: Self

pub const SOUTH: Self

Bottom face button (e.g. Xbox A button)

pub const EAST: Self

Right face button (e.g. Xbox B button)

pub const WEST: Self

Left face button (e.g. Xbox X button)

pub const NORTH: Self

Top face button (e.g. Xbox Y button)

pub const BACK: Self

pub const GUIDE: Self

pub const START: Self

pub const LEFT_STICK: Self

pub const RIGHT_STICK: Self

pub const LEFT_SHOULDER: Self

pub const RIGHT_SHOULDER: Self

pub const DPAD_UP: Self

pub const DPAD_DOWN: Self

pub const DPAD_LEFT: Self

pub const DPAD_RIGHT: Self

pub const MISC1: Self

Additional button (e.g. Xbox Series X share button, PS5 microphone button, Nintendo Switch Pro capture button, Amazon Luna microphone button, Google Stadia capture button)

pub const RIGHT_PADDLE1: Self

Upper or primary paddle, under your right hand (e.g. Xbox Elite paddle P1)

pub const LEFT_PADDLE1: Self

Upper or primary paddle, under your left hand (e.g. Xbox Elite paddle P3)

pub const RIGHT_PADDLE2: Self

Lower or secondary paddle, under your right hand (e.g. Xbox Elite paddle P2)

pub const LEFT_PADDLE2: Self

Lower or secondary paddle, under your left hand (e.g. Xbox Elite paddle P4)

pub const TOUCHPAD: Self

PS4/PS5 touchpad button

pub const MISC2: Self

Additional button

pub const MISC3: Self

Additional button

pub const MISC4: Self

Additional button

pub const MISC5: Self

Additional button

pub const MISC6: Self

Additional button

pub const COUNT: Self