sdl3_sys::gpu
Function SDL_AcquireGPUSwapchainTexture

pub unsafe extern "C" fn SDL_AcquireGPUSwapchainTexture(
    command_buffer: *mut SDL_GPUCommandBuffer,
    window: *mut SDL_Window,
    swapchain_texture: *mut *mut SDL_GPUTexture,
    swapchain_texture_width: *mut Uint32,
    swapchain_texture_height: *mut Uint32,
) -> bool

Acquire a texture to use in presentation.

When a swapchain texture is acquired on a command buffer, it will automatically be submitted for presentation when the command buffer is submitted. The swapchain texture should only be referenced by the command buffer used to acquire it.

This function will fill the swapchain texture handle with NULL if too many frames are in flight. This is not an error.

If you use this function, it is possible to create a situation where many command buffers are allocated while the rendering context waits for the GPU to catch up, which will cause memory usage to grow. You should use SDL_WaitAndAcquireGPUSwapchainTexture() unless you know what you are doing with timing.

The swapchain texture is managed by the implementation and must not be freed by the user. You MUST NOT call this function from any thread other than the one that created the window.

Parameters

command_buffer: a command buffer.

window: a window that has been claimed.

swapchain_texture: a pointer filled in with a swapchain texture handle.

swapchain_texture_width: a pointer filled in with the swapchain texture width, may be NULL.

swapchain_texture_height: a pointer filled in with the swapchain texture height, may be NULL.

Return value

Returns true on success, false on error; call SDL_GetError() for more information.

Thread safety

This function should only be called from the thread that created the window.

Availability

This function is available since SDL 3.2.0.

See also

SDL_ClaimWindowForGPUDevice

SDL_SubmitGPUCommandBuffer

SDL_SubmitGPUCommandBufferAndAcquireFence

SDL_CancelGPUCommandBuffer

SDL_GetWindowSizeInPixels

SDL_WaitForGPUSwapchain

SDL_WaitAndAcquireGPUSwapchainTexture

SDL_SetGPUAllowedFramesInFlight