sdl3_sys::gpu
Function SDL_BeginGPUComputePass

pub unsafe extern "C" fn SDL_BeginGPUComputePass(
    command_buffer: *mut SDL_GPUCommandBuffer,
    storage_texture_bindings: *const SDL_GPUStorageTextureReadWriteBinding,
    num_storage_texture_bindings: Uint32,
    storage_buffer_bindings: *const SDL_GPUStorageBufferReadWriteBinding,
    num_storage_buffer_bindings: Uint32,
) -> *mut SDL_GPUComputePass

Begins a compute pass on a command buffer.

A compute pass is defined by a set of texture subresources and buffers that may be written to by compute pipelines. These textures and buffers must have been created with the COMPUTE_STORAGE_WRITE bit or the COMPUTE_STORAGE_SIMULTANEOUS_READ_WRITE bit. If you do not create a texture with COMPUTE_STORAGE_SIMULTANEOUS_READ_WRITE, you must not read from the texture in the compute pass. All operations related to compute pipelines must take place inside of a compute pass. You must not begin another compute pass, or a render pass or copy pass before ending the compute pass.

A VERY IMPORTANT NOTE - Reads and writes in compute passes are NOT implicitly synchronized. This means you may cause data races by both reading and writing a resource region in a compute pass, or by writing multiple times to a resource region. If your compute work depends on reading the completed output from a previous dispatch, you MUST end the current compute pass and begin a new one before you can safely access the data. Otherwise you will receive unexpected results. Reading and writing a texture in the same compute pass is only supported by specific texture formats. Make sure you check the format support!

Parameters

command_buffer: a command buffer.

storage_texture_bindings: an array of writeable storage texture binding structs.

num_storage_texture_bindings: the number of storage textures to bind from the array.

storage_buffer_bindings: an array of writeable storage buffer binding structs.

num_storage_buffer_bindings: the number of storage buffers to bind from the array.

Return value

Returns a compute pass handle.

Availability

This function is available since SDL 3.2.0.

See also

SDL_EndGPUComputePass