sdl3_sys::gpu
Function SDL_CreateGPUTransferBuffer

pub unsafe extern "C" fn SDL_CreateGPUTransferBuffer(
    device: *mut SDL_GPUDevice,
    createinfo: *const SDL_GPUTransferBufferCreateInfo,
) -> *mut SDL_GPUTransferBuffer

Creates a transfer buffer to be used when uploading to or downloading from graphics resources.

Download buffers can be particularly expensive to create, so it is good practice to reuse them if data will be downloaded regularly.

There are optional properties that can be provided through props. These are the supported properties:

SDL_PROP_GPU_TRANSFERBUFFER_CREATE_NAME_STRING: a name that can be displayed in debugging tools.

Parameters

device: a GPU Context.

createinfo: a struct describing the state of the transfer buffer to create.

Return value

Returns a transfer buffer on success, or NULL on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.

See also

SDL_UploadToGPUBuffer

SDL_DownloadFromGPUBuffer

SDL_UploadToGPUTexture

SDL_DownloadFromGPUTexture

SDL_ReleaseGPUTransferBuffer