sdl3_sys::gpu
Function SDL_DrawGPUPrimitives

pub unsafe extern "C" fn SDL_DrawGPUPrimitives(
    render_pass: *mut SDL_GPURenderPass,
    num_vertices: Uint32,
    num_instances: Uint32,
    first_vertex: Uint32,
    first_instance: Uint32,
)

Draws data using bound graphics state.

You must not call this function before binding a graphics pipeline.

Note that the first_vertex and first_instance parameters are NOT compatible with built-in vertex/instance ID variables in shaders (for example, SV_VertexID); GPU APIs and shader languages do not define these built-in variables consistently, so if your shader depends on them, the only way to keep behavior consistent and portable is to always pass 0 for the correlating parameter in the draw calls.

Parameters

render_pass: a render pass handle.

num_vertices: the number of vertices to draw.

num_instances: the number of instances that will be drawn.

first_vertex: the index of the first vertex to draw.

first_instance: the ID of the first instance to draw.

Availability

This function is available since SDL 3.2.0.