sdl3_sys::gpu
Struct SDL_GPUTextureFormat

#[repr(transparent)]
pub struct SDL_GPUTextureFormat(pub c_int);

Specifies the pixel format of a texture.

Texture format support varies depending on driver, hardware, and usage flags. In general, you should use SDL_GPUTextureSupportsFormat to query if a format is supported before using it. However, there are a few guaranteed formats.

FIXME: Check universal support for 32-bit component formats FIXME: Check universal support for SIMULTANEOUS_READ_WRITE

For SAMPLER usage, the following formats are universally supported:

R8G8B8A8_UNORM

B8G8R8A8_UNORM

R8_UNORM

R8_SNORM

R8G8_UNORM

R8G8_SNORM

R8G8B8A8_SNORM

R16_FLOAT

R16G16_FLOAT

R16G16B16A16_FLOAT

R32_FLOAT

R32G32_FLOAT

R32G32B32A32_FLOAT

R11G11B10_UFLOAT

R8G8B8A8_UNORM_SRGB

B8G8R8A8_UNORM_SRGB

D16_UNORM

For COLOR_TARGET usage, the following formats are universally supported:

R8G8B8A8_UNORM

B8G8R8A8_UNORM

R8_UNORM

R16_FLOAT

R16G16_FLOAT

R16G16B16A16_FLOAT

R32_FLOAT

R32G32_FLOAT

R32G32B32A32_FLOAT

R8_UINT

R8G8_UINT

R8G8B8A8_UINT

R16_UINT

R16G16_UINT

R16G16B16A16_UINT

R8_INT

R8G8_INT

R8G8B8A8_INT

R16_INT

R16G16_INT

R16G16B16A16_INT

R8G8B8A8_UNORM_SRGB

B8G8R8A8_UNORM_SRGB

For STORAGE usages, the following formats are universally supported:

R8G8B8A8_UNORM

R8G8B8A8_SNORM

R16G16B16A16_FLOAT

R32_FLOAT

R32G32_FLOAT

R32G32B32A32_FLOAT

R8G8B8A8_UINT

R16G16B16A16_UINT

R8G8B8A8_INT

R16G16B16A16_INT

For DEPTH_STENCIL_TARGET usage, the following formats are universally supported:

D16_UNORM

Either (but not necessarily both!) D24_UNORM or D32_FLOAT

Either (but not necessarily both!) D24_UNORM_S8_UINT or D32_FLOAT_S8_UINT

Unless D16_UNORM is sufficient for your purposes, always check which of D24/D32 is supported before creating a depth-stencil texture!

Availability

This enum is available since SDL 3.2.0.

See also

SDL_CreateGPUTexture

SDL_GPUTextureSupportsFormat

Known values (sdl3-sys)

Associated constant	Global constant	Description

INVALID	SDL_GPU_TEXTUREFORMAT_INVALID

A8_UNORM	SDL_GPU_TEXTUREFORMAT_A8_UNORM

R8_UNORM	SDL_GPU_TEXTUREFORMAT_R8_UNORM

R8G8_UNORM	SDL_GPU_TEXTUREFORMAT_R8G8_UNORM

R8G8B8A8_UNORM	SDL_GPU_TEXTUREFORMAT_R8G8B8A8_UNORM

R16_UNORM	SDL_GPU_TEXTUREFORMAT_R16_UNORM

R16G16_UNORM	SDL_GPU_TEXTUREFORMAT_R16G16_UNORM

R16G16B16A16_UNORM	SDL_GPU_TEXTUREFORMAT_R16G16B16A16_UNORM

R10G10B10A2_UNORM	SDL_GPU_TEXTUREFORMAT_R10G10B10A2_UNORM

B5G6R5_UNORM	SDL_GPU_TEXTUREFORMAT_B5G6R5_UNORM

B5G5R5A1_UNORM	SDL_GPU_TEXTUREFORMAT_B5G5R5A1_UNORM

B4G4R4A4_UNORM	SDL_GPU_TEXTUREFORMAT_B4G4R4A4_UNORM

B8G8R8A8_UNORM	SDL_GPU_TEXTUREFORMAT_B8G8R8A8_UNORM

BC1_RGBA_UNORM	SDL_GPU_TEXTUREFORMAT_BC1_RGBA_UNORM

BC2_RGBA_UNORM	SDL_GPU_TEXTUREFORMAT_BC2_RGBA_UNORM

BC3_RGBA_UNORM	SDL_GPU_TEXTUREFORMAT_BC3_RGBA_UNORM

BC4_R_UNORM	SDL_GPU_TEXTUREFORMAT_BC4_R_UNORM

BC5_RG_UNORM	SDL_GPU_TEXTUREFORMAT_BC5_RG_UNORM

BC7_RGBA_UNORM	SDL_GPU_TEXTUREFORMAT_BC7_RGBA_UNORM

BC6H_RGB_FLOAT	SDL_GPU_TEXTUREFORMAT_BC6H_RGB_FLOAT

BC6H_RGB_UFLOAT	SDL_GPU_TEXTUREFORMAT_BC6H_RGB_UFLOAT

R8_SNORM	SDL_GPU_TEXTUREFORMAT_R8_SNORM

R8G8_SNORM	SDL_GPU_TEXTUREFORMAT_R8G8_SNORM

R8G8B8A8_SNORM	SDL_GPU_TEXTUREFORMAT_R8G8B8A8_SNORM

R16_SNORM	SDL_GPU_TEXTUREFORMAT_R16_SNORM

R16G16_SNORM	SDL_GPU_TEXTUREFORMAT_R16G16_SNORM

R16G16B16A16_SNORM	SDL_GPU_TEXTUREFORMAT_R16G16B16A16_SNORM

R16_FLOAT	SDL_GPU_TEXTUREFORMAT_R16_FLOAT

R16G16_FLOAT	SDL_GPU_TEXTUREFORMAT_R16G16_FLOAT

R16G16B16A16_FLOAT	SDL_GPU_TEXTUREFORMAT_R16G16B16A16_FLOAT

R32_FLOAT	SDL_GPU_TEXTUREFORMAT_R32_FLOAT

R32G32_FLOAT	SDL_GPU_TEXTUREFORMAT_R32G32_FLOAT

R32G32B32A32_FLOAT	SDL_GPU_TEXTUREFORMAT_R32G32B32A32_FLOAT

R11G11B10_UFLOAT	SDL_GPU_TEXTUREFORMAT_R11G11B10_UFLOAT

R8_UINT	SDL_GPU_TEXTUREFORMAT_R8_UINT

R8G8_UINT	SDL_GPU_TEXTUREFORMAT_R8G8_UINT

R8G8B8A8_UINT	SDL_GPU_TEXTUREFORMAT_R8G8B8A8_UINT

R16_UINT	SDL_GPU_TEXTUREFORMAT_R16_UINT

R16G16_UINT	SDL_GPU_TEXTUREFORMAT_R16G16_UINT

R16G16B16A16_UINT	SDL_GPU_TEXTUREFORMAT_R16G16B16A16_UINT

R32_UINT	SDL_GPU_TEXTUREFORMAT_R32_UINT

R32G32_UINT	SDL_GPU_TEXTUREFORMAT_R32G32_UINT

R32G32B32A32_UINT	SDL_GPU_TEXTUREFORMAT_R32G32B32A32_UINT

R8_INT	SDL_GPU_TEXTUREFORMAT_R8_INT

R8G8_INT	SDL_GPU_TEXTUREFORMAT_R8G8_INT

R8G8B8A8_INT	SDL_GPU_TEXTUREFORMAT_R8G8B8A8_INT

R16_INT	SDL_GPU_TEXTUREFORMAT_R16_INT

R16G16_INT	SDL_GPU_TEXTUREFORMAT_R16G16_INT

R16G16B16A16_INT	SDL_GPU_TEXTUREFORMAT_R16G16B16A16_INT

R32_INT	SDL_GPU_TEXTUREFORMAT_R32_INT

R32G32_INT	SDL_GPU_TEXTUREFORMAT_R32G32_INT

R32G32B32A32_INT	SDL_GPU_TEXTUREFORMAT_R32G32B32A32_INT

R8G8B8A8_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_R8G8B8A8_UNORM_SRGB

B8G8R8A8_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_B8G8R8A8_UNORM_SRGB

BC1_RGBA_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_BC1_RGBA_UNORM_SRGB

BC2_RGBA_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_BC2_RGBA_UNORM_SRGB

BC3_RGBA_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_BC3_RGBA_UNORM_SRGB

BC7_RGBA_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_BC7_RGBA_UNORM_SRGB

D16_UNORM	SDL_GPU_TEXTUREFORMAT_D16_UNORM

D24_UNORM	SDL_GPU_TEXTUREFORMAT_D24_UNORM

D32_FLOAT	SDL_GPU_TEXTUREFORMAT_D32_FLOAT

D24_UNORM_S8_UINT	SDL_GPU_TEXTUREFORMAT_D24_UNORM_S8_UINT

D32_FLOAT_S8_UINT	SDL_GPU_TEXTUREFORMAT_D32_FLOAT_S8_UINT

ASTC_4x4_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_4x4_UNORM

ASTC_5x4_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_5x4_UNORM

ASTC_5x5_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_5x5_UNORM

ASTC_6x5_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_6x5_UNORM

ASTC_6x6_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_6x6_UNORM

ASTC_8x5_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_8x5_UNORM

ASTC_8x6_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_8x6_UNORM

ASTC_8x8_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_8x8_UNORM

ASTC_10x5_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_10x5_UNORM

ASTC_10x6_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_10x6_UNORM

ASTC_10x8_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_10x8_UNORM

ASTC_10x10_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_10x10_UNORM

ASTC_12x10_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_12x10_UNORM

ASTC_12x12_UNORM	SDL_GPU_TEXTUREFORMAT_ASTC_12x12_UNORM

ASTC_4x4_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_4x4_UNORM_SRGB

ASTC_5x4_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_5x4_UNORM_SRGB

ASTC_5x5_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_5x5_UNORM_SRGB

ASTC_6x5_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_6x5_UNORM_SRGB

ASTC_6x6_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_6x6_UNORM_SRGB

ASTC_8x5_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_8x5_UNORM_SRGB

ASTC_8x6_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_8x6_UNORM_SRGB

ASTC_8x8_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_8x8_UNORM_SRGB

ASTC_10x5_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_10x5_UNORM_SRGB

ASTC_10x6_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_10x6_UNORM_SRGB

ASTC_10x8_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_10x8_UNORM_SRGB

ASTC_10x10_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_10x10_UNORM_SRGB

ASTC_12x10_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_12x10_UNORM_SRGB

ASTC_12x12_UNORM_SRGB	SDL_GPU_TEXTUREFORMAT_ASTC_12x12_UNORM_SRGB

ASTC_4x4_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_4x4_FLOAT

ASTC_5x4_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_5x4_FLOAT

ASTC_5x5_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_5x5_FLOAT

ASTC_6x5_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_6x5_FLOAT

ASTC_6x6_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_6x6_FLOAT

ASTC_8x5_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_8x5_FLOAT

ASTC_8x6_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_8x6_FLOAT

ASTC_8x8_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_8x8_FLOAT

ASTC_10x5_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_10x5_FLOAT

ASTC_10x6_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_10x6_FLOAT

ASTC_10x8_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_10x8_FLOAT

ASTC_10x10_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_10x10_FLOAT

ASTC_12x10_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_12x10_FLOAT

ASTC_12x12_FLOAT	SDL_GPU_TEXTUREFORMAT_ASTC_12x12_FLOAT

Tuple Fields
0: c_int
Implementations

impl SDL_GPUTextureFormat

pub const INVALID: Self

pub const A8_UNORM: Self

pub const R8_UNORM: Self

pub const R8G8_UNORM: Self

pub const R8G8B8A8_UNORM: Self

pub const R16_UNORM: Self

pub const R16G16_UNORM: Self

pub const R16G16B16A16_UNORM: Self

pub const R10G10B10A2_UNORM: Self

pub const B5G6R5_UNORM: Self

pub const B5G5R5A1_UNORM: Self

pub const B4G4R4A4_UNORM: Self

pub const B8G8R8A8_UNORM: Self

pub const BC1_RGBA_UNORM: Self

pub const BC2_RGBA_UNORM: Self

pub const BC3_RGBA_UNORM: Self

pub const BC4_R_UNORM: Self

pub const BC5_RG_UNORM: Self

pub const BC7_RGBA_UNORM: Self

pub const BC6H_RGB_FLOAT: Self

pub const BC6H_RGB_UFLOAT: Self

pub const R8_SNORM: Self

pub const R8G8_SNORM: Self

pub const R8G8B8A8_SNORM: Self

pub const R16_SNORM: Self

pub const R16G16_SNORM: Self

pub const R16G16B16A16_SNORM: Self

pub const R16_FLOAT: Self

pub const R16G16_FLOAT: Self

pub const R16G16B16A16_FLOAT: Self

pub const R32_FLOAT: Self

pub const R32G32_FLOAT: Self

pub const R32G32B32A32_FLOAT: Self

pub const R11G11B10_UFLOAT: Self

pub const R8_UINT: Self

pub const R8G8_UINT: Self

pub const R8G8B8A8_UINT: Self

pub const R16_UINT: Self

pub const R16G16_UINT: Self

pub const R16G16B16A16_UINT: Self

pub const R32_UINT: Self

pub const R32G32_UINT: Self

pub const R32G32B32A32_UINT: Self

pub const R8_INT: Self

pub const R8G8_INT: Self

pub const R8G8B8A8_INT: Self

pub const R16_INT: Self

pub const R16G16_INT: Self

pub const R16G16B16A16_INT: Self

pub const R32_INT: Self

pub const R32G32_INT: Self

pub const R32G32B32A32_INT: Self

pub const R8G8B8A8_UNORM_SRGB: Self

pub const B8G8R8A8_UNORM_SRGB: Self

pub const BC1_RGBA_UNORM_SRGB: Self

pub const BC2_RGBA_UNORM_SRGB: Self

pub const BC3_RGBA_UNORM_SRGB: Self

pub const BC7_RGBA_UNORM_SRGB: Self

pub const D16_UNORM: Self

pub const D24_UNORM: Self

pub const D32_FLOAT: Self

pub const D24_UNORM_S8_UINT: Self

pub const D32_FLOAT_S8_UINT: Self

pub const ASTC_4x4_UNORM: Self

pub const ASTC_5x4_UNORM: Self

pub const ASTC_5x5_UNORM: Self

pub const ASTC_6x5_UNORM: Self

pub const ASTC_6x6_UNORM: Self

pub const ASTC_8x5_UNORM: Self

pub const ASTC_8x6_UNORM: Self

pub const ASTC_8x8_UNORM: Self

pub const ASTC_10x5_UNORM: Self

pub const ASTC_10x6_UNORM: Self

pub const ASTC_10x8_UNORM: Self

pub const ASTC_10x10_UNORM: Self

pub const ASTC_12x10_UNORM: Self

pub const ASTC_12x12_UNORM: Self

pub const ASTC_4x4_UNORM_SRGB: Self

pub const ASTC_5x4_UNORM_SRGB: Self

pub const ASTC_5x5_UNORM_SRGB: Self

pub const ASTC_6x5_UNORM_SRGB: Self

pub const ASTC_6x6_UNORM_SRGB: Self

pub const ASTC_8x5_UNORM_SRGB: Self

pub const ASTC_8x6_UNORM_SRGB: Self

pub const ASTC_8x8_UNORM_SRGB: Self

pub const ASTC_10x5_UNORM_SRGB: Self

pub const ASTC_10x6_UNORM_SRGB: Self

pub const ASTC_10x8_UNORM_SRGB: Self

pub const ASTC_10x10_UNORM_SRGB: Self

pub const ASTC_12x10_UNORM_SRGB: Self

pub const ASTC_12x12_UNORM_SRGB: Self

pub const ASTC_4x4_FLOAT: Self

pub const ASTC_5x4_FLOAT: Self

pub const ASTC_5x5_FLOAT: Self

pub const ASTC_6x5_FLOAT: Self

pub const ASTC_6x6_FLOAT: Self

pub const ASTC_8x5_FLOAT: Self

pub const ASTC_8x6_FLOAT: Self

pub const ASTC_8x8_FLOAT: Self

pub const ASTC_10x5_FLOAT: Self

pub const ASTC_10x6_FLOAT: Self

pub const ASTC_10x8_FLOAT: Self

pub const ASTC_10x10_FLOAT: Self

pub const ASTC_12x10_FLOAT: Self

pub const ASTC_12x12_FLOAT: Self