sdl3_sys::gpu
Type Alias SDL_GPUBufferUsageFlags

pub type SDL_GPUBufferUsageFlags = Uint32;

Specifies how a buffer is intended to be used by the client.

A buffer must have at least one usage flag. Note that some usage flag combinations are invalid.

Unlike textures, READ | WRITE can be used for simultaneous read-write usage. The same data synchronization concerns as textures apply.

If you use a STORAGE flag, the data in the buffer must respect std140 layout conventions. In practical terms this means you must ensure that vec3 and vec4 fields are 16-byte aligned.

Availability

This datatype is available since SDL 3.2.0.

See also

SDL_CreateGPUBuffer

Known values (sdl3-sys)

Constant	Description

SDL_GPU_BUFFERUSAGE_VERTEX	Buffer is a vertex buffer.

SDL_GPU_BUFFERUSAGE_INDEX	Buffer is an index buffer.

SDL_GPU_BUFFERUSAGE_INDIRECT	Buffer is an indirect buffer.

SDL_GPU_BUFFERUSAGE_GRAPHICS_STORAGE_READ	Buffer supports storage reads in graphics stages.

SDL_GPU_BUFFERUSAGE_COMPUTE_STORAGE_READ	Buffer supports storage reads in the compute stage.

SDL_GPU_BUFFERUSAGE_COMPUTE_STORAGE_WRITE	Buffer supports storage writes in the compute stage.