sdl3_sys::gpu
Type Alias SDL_GPUTextureUsageFlags

pub type SDL_GPUTextureUsageFlags = Uint32;

Specifies how a texture is intended to be used by the client.

A texture must have at least one usage flag. Note that some usage flag combinations are invalid.

With regards to compute storage usage, READ | WRITE means that you can have shader A that only writes into the texture and shader B that only reads from the texture and bind the same texture to either shader respectively. SIMULTANEOUS means that you can do reads and writes within the same shader or compute pass. It also implies that atomic ops can be used, since those are read-modify-write operations. If you use SIMULTANEOUS, you are responsible for avoiding data races, as there is no data synchronization within a compute pass. Note that SIMULTANEOUS usage is only supported by a limited number of texture formats.

Availability

This datatype is available since SDL 3.2.0.

See also

SDL_CreateGPUTexture

Known values (sdl3-sys)

Constant	Description

SDL_GPU_TEXTUREUSAGE_SAMPLER	Texture supports sampling.

SDL_GPU_TEXTUREUSAGE_COLOR_TARGET	Texture is a color render target.

SDL_GPU_TEXTUREUSAGE_DEPTH_STENCIL_TARGET	Texture is a depth stencil target.

SDL_GPU_TEXTUREUSAGE_GRAPHICS_STORAGE_READ	Texture supports storage reads in graphics stages.

SDL_GPU_TEXTUREUSAGE_COMPUTE_STORAGE_READ	Texture supports storage reads in the compute stage.

SDL_GPU_TEXTUREUSAGE_COMPUTE_STORAGE_WRITE	Texture supports storage writes in the compute stage.

SDL_GPU_TEXTUREUSAGE_COMPUTE_STORAGE_SIMULTANEOUS_READ_WRITE	Texture supports reads and writes in the same compute shader. This is NOT equivalent to READ