sdl3_sys::haptic
Function SDL_OpenHaptic

pub unsafe extern "C" fn SDL_OpenHaptic(
    instance_id: SDL_HapticID,
) -> *mut SDL_Haptic

Open a haptic device for use.

The index passed as an argument refers to the N’th haptic device on this system.

When opening a haptic device, its gain will be set to maximum and autocenter will be disabled. To modify these values use SDL_SetHapticGain() and SDL_SetHapticAutocenter().

Parameters

instance_id: the haptic device instance ID.

Return value

Returns the device identifier or NULL on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.

See also

SDL_CloseHaptic

SDL_GetHaptics

SDL_OpenHapticFromJoystick

SDL_OpenHapticFromMouse

SDL_SetHapticAutocenter

SDL_SetHapticGain