sdl3_sys::haptic
Function SDL_PlayHapticRumble

pub unsafe extern "C" fn SDL_PlayHapticRumble(
    haptic: *mut SDL_Haptic,
    strength: c_float,
    length: Uint32,
) -> bool

Run a simple rumble effect on a haptic device.

Parameters

haptic: the haptic device to play the rumble effect on.

strength: strength of the rumble to play as a 0-1 float value.

length: length of the rumble to play in milliseconds.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.

See also

SDL_InitHapticRumble

SDL_StopHapticRumble