sdl3_sys::hidapi
Function SDL_hid_device_change_count

pub unsafe extern "C" fn SDL_hid_device_change_count() -> Uint32

Check to see if devices may have been added or removed.

Enumerating the HID devices is an expensive operation, so you can call this to see if there have been any system device changes since the last call to this function. A change in the counter returned doesn’t necessarily mean that anything has changed, but you can call SDL_hid_enumerate() to get an updated device list.

Calling this function for the first time may cause a thread or other system resource to be allocated to track device change notifications.

Return value

Returns a change counter that is incremented with each potential device change, or 0 if device change detection isn’t available.

Availability

This function is available since SDL 3.2.0.

See also

SDL_hid_enumerate