sdl3_sys::hidapi
Function SDL_hid_get_feature_report

pub unsafe extern "C" fn SDL_hid_get_feature_report(
    dev: *mut SDL_hid_device,
    data: *mut c_uchar,
    length: usize,
) -> c_int

Get a feature report from a HID device.

Set the first byte of data to the Report ID of the report to be read. Make sure to allow space for this extra byte in data. Upon return, the first byte will still contain the Report ID, and the report data will start in data[1].

Parameters

dev: a device handle returned from SDL_hid_open().

data: a buffer to put the read data into, including the Report ID. Set the first byte of data to the Report ID of the report to be read, or set it to zero if your device does not use numbered reports.

length: the number of bytes to read, including an extra byte for the report ID. The buffer can be longer than the actual report.

Return value

Returns the number of bytes read plus one for the report ID (which is still in the first byte), or -1 on on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.