sdl3_sys::hidapi
Function SDL_hid_send_feature_report

pub unsafe extern "C" fn SDL_hid_send_feature_report(
    dev: *mut SDL_hid_device,
    data: *const c_uchar,
    length: usize,
) -> c_int

Send a Feature report to the device.

Feature reports are sent over the Control endpoint as a Set_Report transfer. The first byte of data must contain the Report ID. For devices which only support a single report, this must be set to 0x0. The remaining bytes contain the report data. Since the Report ID is mandatory, calls to SDL_hid_send_feature_report() will always contain one more byte than the report contains. For example, if a hid report is 16 bytes long, 17 bytes must be passed to SDL_hid_send_feature_report(): the Report ID (or 0x0, for devices which do not use numbered reports), followed by the report data (16 bytes). In this example, the length passed in would be 17.

Parameters

dev: a device handle returned from SDL_hid_open().

data: the data to send, including the report number as the first byte.

length: the length in bytes of the data to send, including the report number.

Return value

Returns the actual number of bytes written and -1 on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.