sdl3_sys
Module hints

This file contains functions to set and get configuration hints, as well as listing each of them alphabetically.

The convention for naming hints is SDL_HINT_X, where “SDL_X” is the environment variable that can be used to override the default.

In general these hints are just that - they may or may not be supported or applicable on any given platform, but they provide a way for an application or user to give the library a hint as to how they would like the library to work.

Structs

SDL_HintPriority

An enumeration of hint priorities.

Constants

SDL_HINT_ALLOW_ALT_TAB_WHILE_GRABBED

Specify the behavior of Alt+Tab while the keyboard is grabbed.

SDL_HINT_ANDROID_ALLOW_RECREATE_ACTIVITY

A variable to control whether the SDL activity is allowed to be re-created.

SDL_HINT_ANDROID_BLOCK_ON_PAUSE

A variable to control whether the event loop will block itself when the app is paused.

SDL_HINT_ANDROID_LOW_LATENCY_AUDIO

A variable to control whether low latency audio should be enabled.

SDL_HINT_ANDROID_TRAP_BACK_BUTTON

A variable to control whether we trap the Android back button to handle it manually.

SDL_HINT_APPLE_TV_CONTROLLER_UI_EVENTS

A variable controlling whether controllers used with the Apple TV generate UI events.

SDL_HINT_APPLE_TV_REMOTE_ALLOW_ROTATION

A variable controlling whether the Apple TV remote’s joystick axes will automatically match the rotation of the remote.

SDL_HINT_APP_ID

A variable setting the app ID string.

SDL_HINT_APP_NAME

A variable setting the application name.

SDL_HINT_ASSERT

A variable controlling response to SDL_assert failures.

SDL_HINT_AUDIO_ALSA_DEFAULT_DEVICE

Specify the default ALSA audio device name.

SDL_HINT_AUDIO_ALSA_DEFAULT_PLAYBACK_DEVICE

Specify the default ALSA audio playback device name.

SDL_HINT_AUDIO_ALSA_DEFAULT_RECORDING_DEVICE

Specify the default ALSA audio recording device name.

SDL_HINT_AUDIO_CATEGORY

A variable controlling the audio category on iOS and macOS.

SDL_HINT_AUDIO_CHANNELS

A variable controlling the default audio channel count.

SDL_HINT_AUDIO_DEVICE_APP_ICON_NAME

Specify an application icon name for an audio device.

SDL_HINT_AUDIO_DEVICE_SAMPLE_FRAMES

A variable controlling device buffer size.

SDL_HINT_AUDIO_DEVICE_STREAM_NAME

Specify an audio stream name for an audio device.

SDL_HINT_AUDIO_DEVICE_STREAM_ROLE

Specify an application role for an audio device.

SDL_HINT_AUDIO_DISK_INPUT_FILE

Specify the input file when recording audio using the disk audio driver.

SDL_HINT_AUDIO_DISK_OUTPUT_FILE

Specify the output file when playing audio using the disk audio driver.

SDL_HINT_AUDIO_DISK_TIMESCALE

A variable controlling the audio rate when using the disk audio driver.

SDL_HINT_AUDIO_DRIVER

A variable that specifies an audio backend to use.

SDL_HINT_AUDIO_DUMMY_TIMESCALE

A variable controlling the audio rate when using the dummy audio driver.

SDL_HINT_AUDIO_FORMAT

A variable controlling the default audio format.

SDL_HINT_AUDIO_FREQUENCY

A variable controlling the default audio frequency.

SDL_HINT_AUDIO_INCLUDE_MONITORS

A variable that causes SDL to not ignore audio “monitors”.

SDL_HINT_AUTO_UPDATE_JOYSTICKS

A variable controlling whether SDL updates joystick state when getting input events.

SDL_HINT_AUTO_UPDATE_SENSORS

A variable controlling whether SDL updates sensor state when getting input events.

SDL_HINT_BMP_SAVE_LEGACY_FORMAT

Prevent SDL from using version 4 of the bitmap header when saving BMPs.

SDL_HINT_CAMERA_DRIVER

A variable that decides what camera backend to use.

SDL_HINT_CPU_FEATURE_MASK

A variable that limits what CPU features are available.

SDL_HINT_DEFAULT

SDL_HINT_DISPLAY_USABLE_BOUNDS

Override for SDL_GetDisplayUsableBounds().

SDL_HINT_EGL_LIBRARY

Specify the EGL library to load.

SDL_HINT_EMSCRIPTEN_ASYNCIFY

Disable giving back control to the browser automatically when running with asyncify.

SDL_HINT_EMSCRIPTEN_CANVAS_SELECTOR

Specify the CSS selector used for the “default” window/canvas.

SDL_HINT_EMSCRIPTEN_KEYBOARD_ELEMENT

Override the binding element for keyboard inputs for Emscripten builds.

SDL_HINT_ENABLE_SCREEN_KEYBOARD

A variable that controls whether the on-screen keyboard should be shown when text input is active.

SDL_HINT_EVDEV_DEVICES

A variable containing a list of evdev devices to use if udev is not available.

SDL_HINT_EVENT_LOGGING

A variable controlling verbosity of the logging of SDL events pushed onto the internal queue.

SDL_HINT_FILE_DIALOG_DRIVER

A variable that specifies a dialog backend to use.

SDL_HINT_FORCE_RAISEWINDOW

A variable controlling whether raising the window should be done more forcefully.

SDL_HINT_FRAMEBUFFER_ACCELERATION

A variable controlling how 3D acceleration is used to accelerate the SDL screen surface.

SDL_HINT_GAMECONTROLLERCONFIG

A variable that lets you manually hint extra gamecontroller db entries.

SDL_HINT_GAMECONTROLLERCONFIG_FILE

A variable that lets you provide a file with extra gamecontroller db entries.

SDL_HINT_GAMECONTROLLERTYPE

A variable that overrides the automatic controller type detection.

SDL_HINT_GAMECONTROLLER_IGNORE_DEVICES

A variable containing a list of devices to skip when scanning for game controllers.

SDL_HINT_GAMECONTROLLER_IGNORE_DEVICES_EXCEPT

If set, all devices will be skipped when scanning for game controllers except for the ones listed in this variable.

SDL_HINT_GAMECONTROLLER_SENSOR_FUSION

A variable that controls whether the device’s built-in accelerometer and gyro should be used as sensors for gamepads.

SDL_HINT_GDK_TEXTINPUT_DEFAULT_TEXT

This variable sets the default text of the TextInput window on GDK platforms.

SDL_HINT_GDK_TEXTINPUT_DESCRIPTION

This variable sets the description of the TextInput window on GDK platforms.

SDL_HINT_GDK_TEXTINPUT_MAX_LENGTH

This variable sets the maximum input length of the TextInput window on GDK platforms.

SDL_HINT_GDK_TEXTINPUT_SCOPE

This variable sets the input scope of the TextInput window on GDK platforms.

SDL_HINT_GDK_TEXTINPUT_TITLE

This variable sets the title of the TextInput window on GDK platforms.

SDL_HINT_GPU_DRIVER

A variable that specifies a GPU backend to use.

SDL_HINT_HIDAPI_ENUMERATE_ONLY_CONTROLLERS

A variable to control whether SDL_hid_enumerate() enumerates all HID devices or only controllers.

SDL_HINT_HIDAPI_IGNORE_DEVICES

A variable containing a list of devices to ignore in SDL_hid_enumerate().

SDL_HINT_HIDAPI_LIBUSB

A variable to control whether HIDAPI uses libusb for device access.

SDL_HINT_HIDAPI_LIBUSB_WHITELIST

A variable to control whether HIDAPI uses libusb only for whitelisted devices.

SDL_HINT_HIDAPI_UDEV

A variable to control whether HIDAPI uses udev for device detection.

SDL_HINT_IME_IMPLEMENTED_UI

A variable describing what IME UI elements the application can display.

SDL_HINT_IOS_HIDE_HOME_INDICATOR

A variable controlling whether the home indicator bar on iPhone X should be hidden.

SDL_HINT_JOYSTICK_ALLOW_BACKGROUND_EVENTS

A variable that lets you enable joystick (and gamecontroller) events even when your app is in the background.

SDL_HINT_JOYSTICK_ARCADESTICK_DEVICES

A variable containing a list of arcade stick style controllers.

SDL_HINT_JOYSTICK_ARCADESTICK_DEVICES_EXCLUDED

A variable containing a list of devices that are not arcade stick style controllers.

SDL_HINT_JOYSTICK_BLACKLIST_DEVICES

A variable containing a list of devices that should not be considered joysticks.

SDL_HINT_JOYSTICK_BLACKLIST_DEVICES_EXCLUDED

A variable containing a list of devices that should be considered joysticks.

SDL_HINT_JOYSTICK_DEVICE

A variable containing a comma separated list of devices to open as joysticks.

SDL_HINT_JOYSTICK_DIRECTINPUT

A variable controlling whether DirectInput should be used for controllers.

SDL_HINT_JOYSTICK_ENHANCED_REPORTS

A variable controlling whether enhanced reports should be used for controllers when using the HIDAPI driver.

SDL_HINT_JOYSTICK_FLIGHTSTICK_DEVICES

A variable containing a list of flightstick style controllers.

SDL_HINT_JOYSTICK_FLIGHTSTICK_DEVICES_EXCLUDED

A variable containing a list of devices that are not flightstick style controllers.

SDL_HINT_JOYSTICK_GAMECUBE_DEVICES

A variable containing a list of devices known to have a GameCube form factor.

SDL_HINT_JOYSTICK_GAMECUBE_DEVICES_EXCLUDED

A variable containing a list of devices known not to have a GameCube form factor.

SDL_HINT_JOYSTICK_GAMEINPUT

A variable controlling whether GameInput should be used for controller handling on Windows.

SDL_HINT_JOYSTICK_HIDAPI

A variable controlling whether the HIDAPI joystick drivers should be used.

SDL_HINT_JOYSTICK_HIDAPI_COMBINE_JOY_CONS

A variable controlling whether Nintendo Switch Joy-Con controllers will be combined into a single Pro-like controller when using the HIDAPI driver.

SDL_HINT_JOYSTICK_HIDAPI_GAMECUBE

A variable controlling whether the HIDAPI driver for Nintendo GameCube controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_GAMECUBE_RUMBLE_BRAKE

A variable controlling whether rumble is used to implement the GameCube controller’s 3 rumble modes, Stop(0), Rumble(1), and StopHard(2).

SDL_HINT_JOYSTICK_HIDAPI_JOYCON_HOME_LED

A variable controlling whether the Home button LED should be turned on when a Nintendo Switch Joy-Con controller is opened.

SDL_HINT_JOYSTICK_HIDAPI_JOY_CONS

A variable controlling whether the HIDAPI driver for Nintendo Switch Joy-Cons should be used.

SDL_HINT_JOYSTICK_HIDAPI_LUNA

A variable controlling whether the HIDAPI driver for Amazon Luna controllers connected via Bluetooth should be used.

SDL_HINT_JOYSTICK_HIDAPI_NINTENDO_CLASSIC

A variable controlling whether the HIDAPI driver for Nintendo Online classic controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_PS3

A variable controlling whether the HIDAPI driver for PS3 controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_PS4

A variable controlling whether the HIDAPI driver for PS4 controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_PS5

A variable controlling whether the HIDAPI driver for PS5 controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_PS3_SIXAXIS_DRIVER

A variable controlling whether the Sony driver (sixaxis.sys) for PS3 controllers (Sixaxis/DualShock 3) should be used.

SDL_HINT_JOYSTICK_HIDAPI_PS4_REPORT_INTERVAL

A variable controlling the update rate of the PS4 controller over Bluetooth when using the HIDAPI driver.

SDL_HINT_JOYSTICK_HIDAPI_PS5_PLAYER_LED

A variable controlling whether the player LEDs should be lit to indicate which player is associated with a PS5 controller.

SDL_HINT_JOYSTICK_HIDAPI_SHIELD

A variable controlling whether the HIDAPI driver for NVIDIA SHIELD controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_STADIA

A variable controlling whether the HIDAPI driver for Google Stadia controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_STEAM

A variable controlling whether the HIDAPI driver for Bluetooth Steam Controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_STEAMDECK

A variable controlling whether the HIDAPI driver for the Steam Deck builtin controller should be used.

SDL_HINT_JOYSTICK_HIDAPI_STEAM_HOME_LED

A variable controlling whether the Steam button LED should be turned on when a Steam controller is opened.

SDL_HINT_JOYSTICK_HIDAPI_STEAM_HORI

A variable controlling whether the HIDAPI driver for HORI licensed Steam controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_SWITCH

A variable controlling whether the HIDAPI driver for Nintendo Switch controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_SWITCH_HOME_LED

A variable controlling whether the Home button LED should be turned on when a Nintendo Switch Pro controller is opened.

SDL_HINT_JOYSTICK_HIDAPI_SWITCH_PLAYER_LED

A variable controlling whether the player LEDs should be lit to indicate which player is associated with a Nintendo Switch controller.

SDL_HINT_JOYSTICK_HIDAPI_VERTICAL_JOY_CONS

A variable controlling whether Nintendo Switch Joy-Con controllers will be in vertical mode when using the HIDAPI driver.

SDL_HINT_JOYSTICK_HIDAPI_WII

A variable controlling whether the HIDAPI driver for Nintendo Wii and Wii U controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_WII_PLAYER_LED

A variable controlling whether the player LEDs should be lit to indicate which player is associated with a Wii controller.

SDL_HINT_JOYSTICK_HIDAPI_XBOX

A variable controlling whether the HIDAPI driver for XBox controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_XBOX_360

A variable controlling whether the HIDAPI driver for XBox 360 controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_XBOX_360_PLAYER_LED

A variable controlling whether the player LEDs should be lit to indicate which player is associated with an Xbox 360 controller.

SDL_HINT_JOYSTICK_HIDAPI_XBOX_360_WIRELESS

A variable controlling whether the HIDAPI driver for XBox 360 wireless controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_XBOX_ONE

A variable controlling whether the HIDAPI driver for XBox One controllers should be used.

SDL_HINT_JOYSTICK_HIDAPI_XBOX_ONE_HOME_LED

A variable controlling whether the Home button LED should be turned on when an Xbox One controller is opened.

SDL_HINT_JOYSTICK_IOKIT

A variable controlling whether IOKit should be used for controller handling.

SDL_HINT_JOYSTICK_LINUX_CLASSIC

A variable controlling whether to use the classic /dev/input/js* joystick interface or the newer /dev/input/event* joystick interface on Linux.

SDL_HINT_JOYSTICK_LINUX_DEADZONES

A variable controlling whether joysticks on Linux adhere to their HID-defined deadzones or return unfiltered values.

SDL_HINT_JOYSTICK_LINUX_DIGITAL_HATS

A variable controlling whether joysticks on Linux will always treat ‘hat’ axis inputs (ABS_HAT0X - ABS_HAT3Y) as 8-way digital hats without checking whether they may be analog.

SDL_HINT_JOYSTICK_LINUX_HAT_DEADZONES

A variable controlling whether digital hats on Linux will apply deadzones to their underlying input axes or use unfiltered values.

SDL_HINT_JOYSTICK_MFI

A variable controlling whether GCController should be used for controller handling.

SDL_HINT_JOYSTICK_RAWINPUT

A variable controlling whether the RAWINPUT joystick drivers should be used for better handling XInput-capable devices.

SDL_HINT_JOYSTICK_RAWINPUT_CORRELATE_XINPUT

A variable controlling whether the RAWINPUT driver should pull correlated data from XInput.

SDL_HINT_JOYSTICK_ROG_CHAKRAM

A variable controlling whether the ROG Chakram mice should show up as joysticks.

SDL_HINT_JOYSTICK_THREAD

A variable controlling whether a separate thread should be used for handling joystick detection and raw input messages on Windows.

SDL_HINT_JOYSTICK_THROTTLE_DEVICES

A variable containing a list of throttle style controllers.

SDL_HINT_JOYSTICK_THROTTLE_DEVICES_EXCLUDED

A variable containing a list of devices that are not throttle style controllers.

SDL_HINT_JOYSTICK_WGI

A variable controlling whether Windows.Gaming.Input should be used for controller handling.

SDL_HINT_JOYSTICK_WHEEL_DEVICES

A variable containing a list of wheel style controllers.

SDL_HINT_JOYSTICK_WHEEL_DEVICES_EXCLUDED

A variable containing a list of devices that are not wheel style controllers.

SDL_HINT_JOYSTICK_ZERO_CENTERED_DEVICES

A variable containing a list of devices known to have all axes centered at zero.

SDL_HINT_KEYCODE_OPTIONS

A variable that controls keycode representation in keyboard events.

SDL_HINT_KMSDRM_DEVICE_INDEX

A variable that controls what KMSDRM device to use.

SDL_HINT_KMSDRM_REQUIRE_DRM_MASTER

A variable that controls whether SDL requires DRM master access in order to initialize the KMSDRM video backend.

SDL_HINT_LOGGING

A variable controlling the default SDL log levels.

SDL_HINT_MAC_BACKGROUND_APP

A variable controlling whether to force the application to become the foreground process when launched on macOS.

SDL_HINT_MAC_CTRL_CLICK_EMULATE_RIGHT_CLICK

A variable that determines whether Ctrl+Click should generate a right-click event on macOS.

SDL_HINT_MAC_OPENGL_ASYNC_DISPATCH

A variable controlling whether dispatching OpenGL context updates should block the dispatching thread until the main thread finishes processing on macOS.

SDL_HINT_MAC_OPTION_AS_ALT

A variable controlling whether the Option key on macOS should be remapped to act as the Alt key.

SDL_HINT_MAC_SCROLL_MOMENTUM

A variable controlling whether SDL_EVENT_MOUSE_WHEEL event values will have momentum on macOS.

SDL_HINT_MAIN_CALLBACK_RATE

Request SDL_AppIterate() be called at a specific rate.

SDL_HINT_MOUSE_AUTO_CAPTURE

A variable controlling whether the mouse is captured while mouse buttons are pressed.

SDL_HINT_MOUSE_DEFAULT_SYSTEM_CURSOR

A variable setting which system cursor to use as the default cursor.

SDL_HINT_MOUSE_DOUBLE_CLICK_RADIUS

A variable setting the double click radius, in pixels.

SDL_HINT_MOUSE_DOUBLE_CLICK_TIME

A variable setting the double click time, in milliseconds.

SDL_HINT_MOUSE_EMULATE_WARP_WITH_RELATIVE

A variable controlling whether warping a hidden mouse cursor will activate relative mouse mode.

SDL_HINT_MOUSE_FOCUS_CLICKTHROUGH

Allow mouse click events when clicking to focus an SDL window.

SDL_HINT_MOUSE_NORMAL_SPEED_SCALE

A variable setting the speed scale for mouse motion, in floating point, when the mouse is not in relative mode.

SDL_HINT_MOUSE_RELATIVE_CURSOR_VISIBLE

A variable controlling whether the hardware cursor stays visible when relative mode is active.

SDL_HINT_MOUSE_RELATIVE_MODE_CENTER

A variable controlling whether relative mouse mode constrains the mouse to the center of the window.

SDL_HINT_MOUSE_RELATIVE_SPEED_SCALE

A variable setting the scale for mouse motion, in floating point, when the mouse is in relative mode.

SDL_HINT_MOUSE_RELATIVE_SYSTEM_SCALE

A variable controlling whether the system mouse acceleration curve is used for relative mouse motion.

SDL_HINT_MOUSE_RELATIVE_WARP_MOTION

A variable controlling whether a motion event should be generated for mouse warping in relative mode.

SDL_HINT_MOUSE_TOUCH_EVENTS

A variable controlling whether mouse events should generate synthetic touch events.

SDL_HINT_MUTE_CONSOLE_KEYBOARD

A variable controlling whether the keyboard should be muted on the console.

SDL_HINT_NORMAL

SDL_HINT_NO_SIGNAL_HANDLERS

Tell SDL not to catch the SIGINT or SIGTERM signals on POSIX platforms.

SDL_HINT_OPENGL_ES_DRIVER

A variable controlling what driver to use for OpenGL ES contexts.

SDL_HINT_OPENGL_LIBRARY

Specify the OpenGL library to load.

SDL_HINT_OPENVR_LIBRARY

Mechanism to specify openvr_api library location

SDL_HINT_ORIENTATIONS

A variable controlling which orientations are allowed on iOS/Android.

SDL_HINT_OVERRIDE

SDL_HINT_PEN_MOUSE_EVENTS

A variable controlling whether pen events should generate synthetic mouse events.

SDL_HINT_PEN_TOUCH_EVENTS

A variable controlling whether pen events should generate synthetic touch events.

SDL_HINT_POLL_SENTINEL

A variable controlling the use of a sentinel event when polling the event queue.

SDL_HINT_PREFERRED_LOCALES

Override for SDL_GetPreferredLocales().

SDL_HINT_QUIT_ON_LAST_WINDOW_CLOSE

A variable that decides whether to send SDL_EVENT_QUIT when closing the last window.

SDL_HINT_RENDER_DIRECT3D11_DEBUG

A variable controlling whether to enable Direct3D 11+’s Debug Layer.

SDL_HINT_RENDER_DIRECT3D_THREADSAFE

A variable controlling whether the Direct3D device is initialized for thread-safe operations.

SDL_HINT_RENDER_DRIVER

A variable specifying which render driver to use.

SDL_HINT_RENDER_GPU_DEBUG

A variable controlling whether to create the GPU device in debug mode.

SDL_HINT_RENDER_GPU_LOW_POWER

A variable controlling whether to prefer a low-power GPU on multi-GPU systems.

SDL_HINT_RENDER_LINE_METHOD

A variable controlling how the 2D render API renders lines.

SDL_HINT_RENDER_METAL_PREFER_LOW_POWER_DEVICE

A variable controlling whether the Metal render driver select low power device over default one.

SDL_HINT_RENDER_VSYNC

A variable controlling whether updates to the SDL screen surface should be synchronized with the vertical refresh, to avoid tearing.

SDL_HINT_RENDER_VULKAN_DEBUG

A variable controlling whether to enable Vulkan Validation Layers.

SDL_HINT_RETURN_KEY_HIDES_IME

A variable to control whether the return key on the soft keyboard should hide the soft keyboard on Android and iOS.

SDL_HINT_ROG_GAMEPAD_MICE

A variable containing a list of ROG gamepad capable mice.

SDL_HINT_ROG_GAMEPAD_MICE_EXCLUDED

A variable containing a list of devices that are not ROG gamepad capable mice.

SDL_HINT_RPI_VIDEO_LAYER

A variable controlling which Dispmanx layer to use on a Raspberry PI.

SDL_HINT_SCREENSAVER_INHIBIT_ACTIVITY_NAME

Specify an “activity name” for screensaver inhibition.

SDL_HINT_SHUTDOWN_DBUS_ON_QUIT

A variable controlling whether SDL calls dbus_shutdown() on quit.

SDL_HINT_STORAGE_TITLE_DRIVER

A variable that specifies a backend to use for title storage.

SDL_HINT_STORAGE_USER_DRIVER

A variable that specifies a backend to use for user storage.

SDL_HINT_THREAD_FORCE_REALTIME_TIME_CRITICAL

Specifies whether SDL_THREAD_PRIORITY_TIME_CRITICAL should be treated as realtime.

SDL_HINT_THREAD_PRIORITY_POLICY

A string specifying additional information to use with SDL_SetCurrentThreadPriority.

SDL_HINT_TIMER_RESOLUTION

A variable that controls the timer resolution, in milliseconds.

SDL_HINT_TOUCH_MOUSE_EVENTS

A variable controlling whether touch events should generate synthetic mouse events.

SDL_HINT_TRACKPAD_IS_TOUCH_ONLY

A variable controlling whether trackpads should be treated as touch devices.

SDL_HINT_TV_REMOTE_AS_JOYSTICK

A variable controlling whether the Android / tvOS remotes should be listed as joystick devices, instead of sending keyboard events.

SDL_HINT_VIDEO_ALLOW_SCREENSAVER

A variable controlling whether the screensaver is enabled.

SDL_HINT_VIDEO_DISPLAY_PRIORITY

A comma separated list containing the names of the displays that SDL should sort to the front of the display list.

SDL_HINT_VIDEO_DOUBLE_BUFFER

Tell the video driver that we only want a double buffer.

SDL_HINT_VIDEO_DRIVER

A variable that specifies a video backend to use.

SDL_HINT_VIDEO_DUMMY_SAVE_FRAMES

A variable controlling whether the dummy video driver saves output frames.

SDL_HINT_VIDEO_EGL_ALLOW_GETDISPLAY_FALLBACK

If eglGetPlatformDisplay fails, fall back to calling eglGetDisplay.

SDL_HINT_VIDEO_FORCE_EGL

A variable controlling whether the OpenGL context should be created with EGL.

SDL_HINT_VIDEO_MAC_FULLSCREEN_MENU_VISIBILITY

A variable that specifies the menu visibility when a window is fullscreen in Spaces on macOS.

SDL_HINT_VIDEO_MAC_FULLSCREEN_SPACES

A variable that specifies the policy for fullscreen Spaces on macOS.

SDL_HINT_VIDEO_MINIMIZE_ON_FOCUS_LOSS

A variable controlling whether fullscreen windows are minimized when they lose focus.

SDL_HINT_VIDEO_OFFSCREEN_SAVE_FRAMES

A variable controlling whether the offscreen video driver saves output frames.

SDL_HINT_VIDEO_SYNC_WINDOW_OPERATIONS

A variable controlling whether all window operations will block until complete.

SDL_HINT_VIDEO_WAYLAND_ALLOW_LIBDECOR

A variable controlling whether the libdecor Wayland backend is allowed to be used.

SDL_HINT_VIDEO_WAYLAND_MODE_EMULATION

A variable controlling whether video mode emulation is enabled under Wayland.

SDL_HINT_VIDEO_WAYLAND_MODE_SCALING

A variable controlling how modes with a non-native aspect ratio are displayed under Wayland.

SDL_HINT_VIDEO_WAYLAND_PREFER_LIBDECOR

A variable controlling whether the libdecor Wayland backend is preferred over native decorations.

SDL_HINT_VIDEO_WAYLAND_SCALE_TO_DISPLAY

A variable forcing non-DPI-aware Wayland windows to output at 1:1 scaling.

SDL_HINT_VIDEO_WIN_D3DCOMPILER

A variable specifying which shader compiler to preload when using the Chrome ANGLE binaries.

SDL_HINT_VIDEO_X11_NET_WM_BYPASS_COMPOSITOR

A variable controlling whether the X11 _NET_WM_BYPASS_COMPOSITOR hint should be used.

SDL_HINT_VIDEO_X11_NET_WM_PING

A variable controlling whether the X11 _NET_WM_PING protocol should be supported.

SDL_HINT_VIDEO_X11_NODIRECTCOLOR

A variable controlling whether SDL uses DirectColor visuals.

SDL_HINT_VIDEO_X11_SCALING_FACTOR

A variable forcing the content scaling factor for X11 displays.

SDL_HINT_VIDEO_X11_VISUALID

A variable forcing the visual ID used for X11 display modes.

SDL_HINT_VIDEO_X11_WINDOW_VISUALID

A variable forcing the visual ID chosen for new X11 windows.

SDL_HINT_VIDEO_X11_XRANDR

A variable controlling whether the X11 XRandR extension should be used.

SDL_HINT_VITA_ENABLE_BACK_TOUCH

A variable controlling whether touch should be enabled on the back panel of the PlayStation Vita.

SDL_HINT_VITA_ENABLE_FRONT_TOUCH

A variable controlling whether touch should be enabled on the front panel of the PlayStation Vita.

SDL_HINT_VITA_MODULE_PATH

A variable controlling the module path on the PlayStation Vita.

SDL_HINT_VITA_PVR_INIT

A variable controlling whether to perform PVR initialization on the PlayStation Vita.

SDL_HINT_VITA_PVR_OPENGL

A variable controlling whether OpenGL should be used instead of OpenGL ES on the PlayStation Vita.

SDL_HINT_VITA_RESOLUTION

A variable overriding the resolution reported on the PlayStation Vita.

SDL_HINT_VITA_TOUCH_MOUSE_DEVICE

A variable controlling which touchpad should generate synthetic mouse events.

SDL_HINT_VULKAN_DISPLAY

A variable overriding the display index used in SDL_Vulkan_CreateSurface()

SDL_HINT_VULKAN_LIBRARY

Specify the Vulkan library to load.

SDL_HINT_WAVE_CHUNK_LIMIT

A variable controlling the maximum number of chunks in a WAVE file.

SDL_HINT_WAVE_FACT_CHUNK

A variable controlling how the fact chunk affects the loading of a WAVE file.

SDL_HINT_WAVE_RIFF_CHUNK_SIZE

A variable controlling how the size of the RIFF chunk affects the loading of a WAVE file.

SDL_HINT_WAVE_TRUNCATION

A variable controlling how a truncated WAVE file is handled.

SDL_HINT_WINDOWS_CLOSE_ON_ALT_F4

A variable controlling whether SDL generates window-close events for Alt+F4 on Windows.

SDL_HINT_WINDOWS_ENABLE_MENU_MNEMONICS

A variable controlling whether menus can be opened with their keyboard shortcut (Alt+mnemonic).

SDL_HINT_WINDOWS_ENABLE_MESSAGELOOP

A variable controlling whether the windows message loop is processed by SDL.

SDL_HINT_WINDOWS_ERASE_BACKGROUND_MODE

A variable controlling whether SDL will clear the window contents when the WM_ERASEBKGND message is received.

SDL_HINT_WINDOWS_FORCE_SEMAPHORE_KERNEL

A variable controlling whether SDL uses Kernel Semaphores on Windows.

SDL_HINT_WINDOWS_GAMEINPUT

A variable controlling whether GameInput is used for raw keyboard and mouse on Windows.

SDL_HINT_WINDOWS_INTRESOURCE_ICON

A variable to specify custom icon resource id from RC file on Windows platform.

SDL_HINT_WINDOWS_INTRESOURCE_ICON_SMALL

A variable to specify custom icon resource id from RC file on Windows platform.

SDL_HINT_WINDOWS_RAW_KEYBOARD

A variable controlling whether raw keyboard events are used on Windows.

SDL_HINT_WINDOWS_USE_D3D9EX

A variable controlling whether SDL uses the D3D9Ex API introduced in Windows Vista, instead of normal D3D9.

SDL_HINT_WINDOW_ACTIVATE_WHEN_RAISED

A variable controlling whether the window is activated when the SDL_RaiseWindow function is called.

SDL_HINT_WINDOW_ACTIVATE_WHEN_SHOWN

A variable controlling whether the window is activated when the SDL_ShowWindow function is called.

SDL_HINT_WINDOW_ALLOW_TOPMOST

If set to “0” then never set the top-most flag on an SDL Window even if the application requests it.

SDL_HINT_WINDOW_FRAME_USABLE_WHILE_CURSOR_HIDDEN

A variable controlling whether the window frame and title bar are interactive when the cursor is hidden.

SDL_HINT_X11_FORCE_OVERRIDE_REDIRECT

A variable controlling whether X11 windows are marked as override-redirect.

SDL_HINT_X11_WINDOW_TYPE

A variable specifying the type of an X11 window.

SDL_HINT_X11_XCB_LIBRARY

Specify the XCB library to load for the X11 driver.

SDL_HINT_XINPUT_ENABLED

A variable controlling whether XInput should be used for controller handling.

Functions

SDL_AddHintCallback⚠

Add a function to watch a particular hint.

SDL_GetHint⚠

Get the value of a hint.

SDL_GetHintBoolean⚠

Get the boolean value of a hint variable.

SDL_RemoveHintCallback⚠

Remove a function watching a particular hint.

SDL_ResetHint⚠

Reset a hint to the default value.

SDL_ResetHints⚠

Reset all hints to the default values.

SDL_SetHint⚠

Set a hint with normal priority.

SDL_SetHintWithPriority⚠

Set a hint with a specific priority.

Type Aliases

SDL_HintCallback

A callback used to send notifications of hint value changes.