sdl3_sys::init
Function SDL_IsMainThread

pub unsafe extern "C" fn SDL_IsMainThread() -> bool

Return whether this is the main thread.

On Apple platforms, the main thread is the thread that runs your program’s main() entry point. On other platforms, the main thread is the one that calls SDL_Init(SDL_INIT_VIDEO), which should usually be the one that runs your program’s main() entry point. If you are using the main callbacks, SDL_AppInit(), SDL_AppIterate(), and SDL_AppQuit() are all called on the main thread.

Return value

Returns true if this thread is the main thread, or false otherwise.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_RunOnMainThread