sdl3_sys::iostream
Function SDL_ReadS32LE

pub unsafe extern "C" fn SDL_ReadS32LE(
    src: *mut SDL_IOStream,
    value: *mut Sint32,
) -> bool

Use this function to read 32 bits of little-endian data from an SDL_IOStream and return in native format.

SDL byteswaps the data only if necessary, so the data returned will be in the native byte order.

This function will return false when the data stream is completely read, and SDL_GetIOStatus() will return SDL_IO_STATUS_EOF. If false is returned and the stream is not at EOF, SDL_GetIOStatus() will return a different error value and SDL_GetError() will offer a human-readable message.

Parameters

src: the stream from which to read data.

value: a pointer filled in with the data read.

Return value

Returns true on successful write or false on failure; call SDL_GetError() for more information.

Thread safety

This function is not thread safe.

Availability

This function is available since SDL 3.2.0.