sdl3_sys::iostream
Function SDL_TellIO

pub unsafe extern "C" fn SDL_TellIO(
    context: *mut SDL_IOStream,
) -> Sint64

Determine the current read/write offset in an SDL_IOStream data stream.

SDL_TellIO is actually a wrapper function that calls the SDL_IOStream’s seek method, with an offset of 0 bytes from SDL_IO_SEEK_CUR, to simplify application development.

Parameters

context: an SDL_IOStream data stream object from which to get the current offset.

Return value

Returns the current offset in the stream, or -1 if the information can not be determined.

Thread safety

This function is not thread safe.

Availability

This function is available since SDL 3.2.0.

See also

SDL_SeekIO