sdl3_sys::keyboard
Function SDL_GetKeyFromScancode

pub unsafe extern "C" fn SDL_GetKeyFromScancode(
    scancode: SDL_Scancode,
    modstate: SDL_Keymod,
    key_event: bool,
) -> SDL_Keycode

Get the key code corresponding to the given scancode according to the current keyboard layout.

If you want to get the keycode as it would be delivered in key events, including options specified in SDL_HINT_KEYCODE_OPTIONS, then you should pass key_event as true. Otherwise this function simply translates the scancode based on the given modifier state.

Parameters

scancode: the desired SDL_Scancode to query.

modstate: the modifier state to use when translating the scancode to a keycode.

key_event: true if the keycode will be used in key events.

Return value

Returns the SDL_Keycode that corresponds to the given SDL_Scancode.

Thread safety

This function is not thread safe.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetKeyName

SDL_GetScancodeFromKey