sdl3_sys::log
Type Alias SDL_LogOutputFunction

pub type SDL_LogOutputFunction = Option<unsafe extern "C" fn(userdata: *mut c_void, category: c_int, priority: SDL_LogPriority, message: *const c_char)>;

The prototype for the log output callback function.

This function is called by SDL when there is new text to be logged. A mutex is held so that this function is never called by more than one thread at once.

Parameters

userdata: what was passed as userdata to SDL_SetLogOutputFunction().

category: the category of the message.

priority: the priority of the message.

message: the message being output.

Availability

This datatype is available since SDL 3.2.0.

Aliased Type

enum SDL_LogOutputFunction {
    None,
    Some(unsafe extern "C" fn(_: *mut c_void, _: i32, _: SDL_LogPriority, _: *const i8)),
}

Variants

None

No value.

Some(unsafe extern "C" fn(_: *mut c_void, _: i32, _: SDL_LogPriority, _: *const i8))

Some value of type T.