sdl3_sys::main
Function SDL_RegisterApp

pub unsafe extern "C" fn SDL_RegisterApp(
    name: *const c_char,
    style: Uint32,
    hInst: *mut c_void,
) -> bool

Register a win32 window class for SDL’s use.

This can be called to set the application window class at startup. It is safe to call this multiple times, as long as every call is eventually paired with a call to SDL_UnregisterApp, but a second registration attempt while a previous registration is still active will be ignored, other than to increment a counter.

Most applications do not need to, and should not, call this directly; SDL will call it when initializing the video subsystem.

Parameters

name: the window class name, in UTF-8 encoding. If NULL, SDL currently uses “SDL_app” but this isn’t guaranteed.

style: the value to use in WNDCLASSEX::style. If name is NULL, SDL currently uses (CS_BYTEALIGNCLIENT | CS_OWNDC) regardless of what is specified here.

hInst: the HINSTANCE to use in WNDCLASSEX::hInstance. If zero, SDL will use GetModuleHandle(NULL) instead.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.