sdl3_sys::mouse
Struct SDL_SystemCursor

#[repr(transparent)]
pub struct SDL_SystemCursor(pub c_int);

Cursor types for SDL_CreateSystemCursor().

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

DEFAULT	SDL_SYSTEM_CURSOR_DEFAULT	Default cursor. Usually an arrow.

TEXT	SDL_SYSTEM_CURSOR_TEXT	Text selection. Usually an I-beam.

WAIT	SDL_SYSTEM_CURSOR_WAIT	Wait. Usually an hourglass or watch or spinning ball.

CROSSHAIR	SDL_SYSTEM_CURSOR_CROSSHAIR	Crosshair.

PROGRESS	SDL_SYSTEM_CURSOR_PROGRESS	Program is busy but still interactive. Usually it’s WAIT with an arrow.

NWSE_RESIZE	SDL_SYSTEM_CURSOR_NWSE_RESIZE	Double arrow pointing northwest and southeast.

NESW_RESIZE	SDL_SYSTEM_CURSOR_NESW_RESIZE	Double arrow pointing northeast and southwest.

EW_RESIZE	SDL_SYSTEM_CURSOR_EW_RESIZE	Double arrow pointing west and east.

NS_RESIZE	SDL_SYSTEM_CURSOR_NS_RESIZE	Double arrow pointing north and south.

MOVE	SDL_SYSTEM_CURSOR_MOVE	Four pointed arrow pointing north, south, east, and west.

NOT_ALLOWED	SDL_SYSTEM_CURSOR_NOT_ALLOWED	Not permitted. Usually a slashed circle or crossbones.

POINTER	SDL_SYSTEM_CURSOR_POINTER	Pointer that indicates a link. Usually a pointing hand.

NW_RESIZE	SDL_SYSTEM_CURSOR_NW_RESIZE	Window resize top-left. This may be a single arrow or a double arrow like NWSE_RESIZE.

N_RESIZE	SDL_SYSTEM_CURSOR_N_RESIZE	Window resize top. May be NS_RESIZE.

NE_RESIZE	SDL_SYSTEM_CURSOR_NE_RESIZE	Window resize top-right. May be NESW_RESIZE.

E_RESIZE	SDL_SYSTEM_CURSOR_E_RESIZE	Window resize right. May be EW_RESIZE.

SE_RESIZE	SDL_SYSTEM_CURSOR_SE_RESIZE	Window resize bottom-right. May be NWSE_RESIZE.

S_RESIZE	SDL_SYSTEM_CURSOR_S_RESIZE	Window resize bottom. May be NS_RESIZE.

SW_RESIZE	SDL_SYSTEM_CURSOR_SW_RESIZE	Window resize bottom-left. May be NESW_RESIZE.

W_RESIZE	SDL_SYSTEM_CURSOR_W_RESIZE	Window resize left. May be EW_RESIZE.

COUNT	SDL_SYSTEM_CURSOR_COUNT

Tuple Fields
0: c_int
Implementations

impl SDL_SystemCursor

pub const DEFAULT: Self

Default cursor. Usually an arrow.

pub const TEXT: Self

Text selection. Usually an I-beam.

pub const WAIT: Self

Wait. Usually an hourglass or watch or spinning ball.

pub const CROSSHAIR: Self

Crosshair.

pub const PROGRESS: Self

Program is busy but still interactive. Usually it’s WAIT with an arrow.

pub const NWSE_RESIZE: Self

Double arrow pointing northwest and southeast.

pub const NESW_RESIZE: Self

Double arrow pointing northeast and southwest.

pub const EW_RESIZE: Self

Double arrow pointing west and east.

pub const NS_RESIZE: Self

Double arrow pointing north and south.

pub const MOVE: Self

Four pointed arrow pointing north, south, east, and west.

pub const NOT_ALLOWED: Self

Not permitted. Usually a slashed circle or crossbones.

pub const POINTER: Self

Pointer that indicates a link. Usually a pointing hand.

pub const NW_RESIZE: Self

Window resize top-left. This may be a single arrow or a double arrow like NWSE_RESIZE.

pub const N_RESIZE: Self

Window resize top. May be NS_RESIZE.

pub const NE_RESIZE: Self

Window resize top-right. May be NESW_RESIZE.

pub const E_RESIZE: Self

Window resize right. May be EW_RESIZE.

pub const SE_RESIZE: Self

Window resize bottom-right. May be NWSE_RESIZE.

pub const S_RESIZE: Self

Window resize bottom. May be NS_RESIZE.

pub const SW_RESIZE: Self

Window resize bottom-left. May be NESW_RESIZE.

pub const W_RESIZE: Self

Window resize left. May be EW_RESIZE.

pub const COUNT: Self