sdl3_sys::pixels
Function SDL_DEFINE_COLORSPACE

pub const fn SDL_DEFINE_COLORSPACE(
    type: SDL_ColorType,
    range: SDL_ColorRange,
    primaries: SDL_ColorPrimaries,
    transfer: SDL_TransferCharacteristics,
    matrix: SDL_MatrixCoefficients,
    chroma: SDL_ChromaLocation,
) -> SDL_Colorspace

A macro for defining custom SDL_Colorspace formats.

For example, defining SDL_COLORSPACE_SRGB looks like this:

SDL_DEFINE_COLORSPACE(SDL_COLOR_TYPE_RGB,
                      SDL_COLOR_RANGE_FULL,
                      SDL_COLOR_PRIMARIES_BT709,
                      SDL_TRANSFER_CHARACTERISTICS_SRGB,
                      SDL_MATRIX_COEFFICIENTS_IDENTITY,
                      SDL_CHROMA_LOCATION_NONE)

Parameters

type: the type of the new format, probably an SDL_ColorType value.

range: the range of the new format, probably a SDL_ColorRange value.

primaries: the primaries of the new format, probably an SDL_ColorPrimaries value.

transfer: the transfer characteristics of the new format, probably an SDL_TransferCharacteristics value.

matrix: the matrix coefficients of the new format, probably an SDL_MatrixCoefficients value.

chroma: the chroma sample location of the new format, probably an SDL_ChromaLocation value.

Return value

Returns a format value in the style of SDL_Colorspace.

Thread safety

It is safe to call this macro from any thread.

Availability

This macro is available since SDL 3.2.0.