sdl3_sys::pixels
Struct SDL_PixelFormat

#[repr(transparent)]
pub struct SDL_PixelFormat(pub c_int);

Pixel format.

SDL’s pixel formats have the following naming convention:

Names with a list of components and a single bit count, such as RGB24 and ABGR32, define a platform-independent encoding into bytes in the order specified. For example, in RGB24 data, each pixel is encoded in 3 bytes (red, green, blue) in that order, and in ABGR32 data, each pixel is encoded in 4 bytes alpha, blue, green, red) in that order. Use these names if the property of a format that is important to you is the order of the bytes in memory or on disk.

Names with a bit count per component, such as ARGB8888 and XRGB1555, are “packed” into an appropriately-sized integer in the platform’s native endianness. For example, ARGB8888 is a sequence of 32-bit integers; in each integer, the most significant bits are alpha, and the least significant bits are blue. On a little-endian CPU such as x86, the least significant bits of each integer are arranged first in memory, but on a big-endian CPU such as s390x, the most significant bits are arranged first. Use these names if the property of a format that is important to you is the meaning of each bit position within a native-endianness integer.

In indexed formats such as INDEX4LSB, each pixel is represented by encoding an index into the palette into the indicated number of bits, with multiple pixels packed into each byte if appropriate. In LSB formats, the first (leftmost) pixel is stored in the least-significant bits of the byte; in MSB formats, it’s stored in the most-significant bits. INDEX8 does not need LSB/MSB variants, because each pixel exactly fills one byte.

The 32-bit byte-array encodings such as RGBA32 are aliases for the appropriate 8888 encoding for the current platform. For example, RGBA32 is an alias for ABGR8888 on little-endian CPUs like x86, or an alias for RGBA8888 on big-endian CPUs.

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

UNKNOWN	SDL_PIXELFORMAT_UNKNOWN

INDEX1LSB	SDL_PIXELFORMAT_INDEX1LSB

INDEX1MSB	SDL_PIXELFORMAT_INDEX1MSB

INDEX2LSB	SDL_PIXELFORMAT_INDEX2LSB

INDEX2MSB	SDL_PIXELFORMAT_INDEX2MSB

INDEX4LSB	SDL_PIXELFORMAT_INDEX4LSB

INDEX4MSB	SDL_PIXELFORMAT_INDEX4MSB

INDEX8	SDL_PIXELFORMAT_INDEX8

RGB332	SDL_PIXELFORMAT_RGB332

XRGB4444	SDL_PIXELFORMAT_XRGB4444

XBGR4444	SDL_PIXELFORMAT_XBGR4444

XRGB1555	SDL_PIXELFORMAT_XRGB1555

XBGR1555	SDL_PIXELFORMAT_XBGR1555

ARGB4444	SDL_PIXELFORMAT_ARGB4444

RGBA4444	SDL_PIXELFORMAT_RGBA4444

ABGR4444	SDL_PIXELFORMAT_ABGR4444

BGRA4444	SDL_PIXELFORMAT_BGRA4444

ARGB1555	SDL_PIXELFORMAT_ARGB1555

RGBA5551	SDL_PIXELFORMAT_RGBA5551

ABGR1555	SDL_PIXELFORMAT_ABGR1555

BGRA5551	SDL_PIXELFORMAT_BGRA5551

RGB565	SDL_PIXELFORMAT_RGB565

BGR565	SDL_PIXELFORMAT_BGR565

RGB24	SDL_PIXELFORMAT_RGB24

BGR24	SDL_PIXELFORMAT_BGR24

XRGB8888	SDL_PIXELFORMAT_XRGB8888

RGBX8888	SDL_PIXELFORMAT_RGBX8888

XBGR8888	SDL_PIXELFORMAT_XBGR8888

BGRX8888	SDL_PIXELFORMAT_BGRX8888

ARGB8888	SDL_PIXELFORMAT_ARGB8888

RGBA8888	SDL_PIXELFORMAT_RGBA8888

ABGR8888	SDL_PIXELFORMAT_ABGR8888

BGRA8888	SDL_PIXELFORMAT_BGRA8888

XRGB2101010	SDL_PIXELFORMAT_XRGB2101010

XBGR2101010	SDL_PIXELFORMAT_XBGR2101010

ARGB2101010	SDL_PIXELFORMAT_ARGB2101010

ABGR2101010	SDL_PIXELFORMAT_ABGR2101010

RGB48	SDL_PIXELFORMAT_RGB48

BGR48	SDL_PIXELFORMAT_BGR48

RGBA64	SDL_PIXELFORMAT_RGBA64

ARGB64	SDL_PIXELFORMAT_ARGB64

BGRA64	SDL_PIXELFORMAT_BGRA64

ABGR64	SDL_PIXELFORMAT_ABGR64

RGB48_FLOAT	SDL_PIXELFORMAT_RGB48_FLOAT

BGR48_FLOAT	SDL_PIXELFORMAT_BGR48_FLOAT

RGBA64_FLOAT	SDL_PIXELFORMAT_RGBA64_FLOAT

ARGB64_FLOAT	SDL_PIXELFORMAT_ARGB64_FLOAT

BGRA64_FLOAT	SDL_PIXELFORMAT_BGRA64_FLOAT

ABGR64_FLOAT	SDL_PIXELFORMAT_ABGR64_FLOAT

RGB96_FLOAT	SDL_PIXELFORMAT_RGB96_FLOAT

BGR96_FLOAT	SDL_PIXELFORMAT_BGR96_FLOAT

RGBA128_FLOAT	SDL_PIXELFORMAT_RGBA128_FLOAT

ARGB128_FLOAT	SDL_PIXELFORMAT_ARGB128_FLOAT

BGRA128_FLOAT	SDL_PIXELFORMAT_BGRA128_FLOAT

ABGR128_FLOAT	SDL_PIXELFORMAT_ABGR128_FLOAT

YV12	SDL_PIXELFORMAT_YV12	Planar mode: Y + V + U (3 planes)

IYUV	SDL_PIXELFORMAT_IYUV	Planar mode: Y + U + V (3 planes)

YUY2	SDL_PIXELFORMAT_YUY2	Packed mode: Y0+U0+Y1+V0 (1 plane)

UYVY	SDL_PIXELFORMAT_UYVY	Packed mode: U0+Y0+V0+Y1 (1 plane)

YVYU	SDL_PIXELFORMAT_YVYU	Packed mode: Y0+V0+Y1+U0 (1 plane)

NV12	SDL_PIXELFORMAT_NV12	Planar mode: Y + U/V interleaved (2 planes)

NV21	SDL_PIXELFORMAT_NV21	Planar mode: Y + V/U interleaved (2 planes)

P010	SDL_PIXELFORMAT_P010	Planar mode: Y + U/V interleaved (2 planes)

EXTERNAL_OES	SDL_PIXELFORMAT_EXTERNAL_OES	Android video texture format

RGBA32	SDL_PIXELFORMAT_RGBA32	(target dependent)

ARGB32	SDL_PIXELFORMAT_ARGB32	(target dependent)

BGRA32	SDL_PIXELFORMAT_BGRA32	(target dependent)

ABGR32	SDL_PIXELFORMAT_ABGR32	(target dependent)

RGBX32	SDL_PIXELFORMAT_RGBX32	(target dependent)

XRGB32	SDL_PIXELFORMAT_XRGB32	(target dependent)

BGRX32	SDL_PIXELFORMAT_BGRX32	(target dependent)

XBGR32	SDL_PIXELFORMAT_XBGR32	(target dependent)

Tuple Fields
0: c_int
Implementations

impl SDL_PixelFormat

pub const UNKNOWN: Self

pub const INDEX1LSB: Self

pub const INDEX1MSB: Self

pub const INDEX2LSB: Self

pub const INDEX2MSB: Self

pub const INDEX4LSB: Self

pub const INDEX4MSB: Self

pub const INDEX8: Self

pub const RGB332: Self

pub const XRGB4444: Self

pub const XBGR4444: Self

pub const XRGB1555: Self

pub const XBGR1555: Self

pub const ARGB4444: Self

pub const RGBA4444: Self

pub const ABGR4444: Self

pub const BGRA4444: Self

pub const ARGB1555: Self

pub const RGBA5551: Self

pub const ABGR1555: Self

pub const BGRA5551: Self

pub const RGB565: Self

pub const BGR565: Self

pub const RGB24: Self

pub const BGR24: Self

pub const XRGB8888: Self

pub const RGBX8888: Self

pub const XBGR8888: Self

pub const BGRX8888: Self

pub const ARGB8888: Self

pub const RGBA8888: Self

pub const ABGR8888: Self

pub const BGRA8888: Self

pub const XRGB2101010: Self

pub const XBGR2101010: Self

pub const ARGB2101010: Self

pub const ABGR2101010: Self

pub const RGB48: Self

pub const BGR48: Self

pub const RGBA64: Self

pub const ARGB64: Self

pub const BGRA64: Self

pub const ABGR64: Self

pub const RGB48_FLOAT: Self

pub const BGR48_FLOAT: Self

pub const RGBA64_FLOAT: Self

pub const ARGB64_FLOAT: Self

pub const BGRA64_FLOAT: Self

pub const ABGR64_FLOAT: Self

pub const RGB96_FLOAT: Self

pub const BGR96_FLOAT: Self

pub const RGBA128_FLOAT: Self

pub const ARGB128_FLOAT: Self

pub const BGRA128_FLOAT: Self

pub const ABGR128_FLOAT: Self

pub const YV12: Self

Planar mode: Y + V + U (3 planes)

pub const IYUV: Self

Planar mode: Y + U + V (3 planes)

pub const YUY2: Self

Packed mode: Y0+U0+Y1+V0 (1 plane)

pub const UYVY: Self

Packed mode: U0+Y0+V0+Y1 (1 plane)

pub const YVYU: Self

Packed mode: Y0+V0+Y1+U0 (1 plane)

pub const NV12: Self

Planar mode: Y + U/V interleaved (2 planes)

pub const NV21: Self

Planar mode: Y + V/U interleaved (2 planes)

pub const P010: Self

Planar mode: Y + U/V interleaved (2 planes)

pub const EXTERNAL_OES: Self

Android video texture format

pub const RGBA32: Self = SDL_PIXELFORMAT_ABGR8888

pub const ARGB32: Self = SDL_PIXELFORMAT_BGRA8888

pub const BGRA32: Self = SDL_PIXELFORMAT_ARGB8888

pub const ABGR32: Self = SDL_PIXELFORMAT_RGBA8888

pub const RGBX32: Self = SDL_PIXELFORMAT_XBGR8888

pub const XRGB32: Self = SDL_PIXELFORMAT_BGRX8888

pub const BGRX32: Self = SDL_PIXELFORMAT_XRGB8888

pub const XBGR32: Self = SDL_PIXELFORMAT_RGBX8888