sdl3_sys
Module power

SDL power management routines.

There is a single function in this category: SDL_GetPowerInfo().

This function is useful for games on the go. This allows an app to know if it’s running on a draining battery, which can be useful if the app wants to reduce processing, or perhaps framerate, to extend the duration of the battery’s charge. Perhaps the app just wants to show a battery meter when fullscreen, or alert the user when the power is getting extremely low, so they can save their game.

Structs

SDL_PowerState

The basic state for the system’s power supply.

Constants

SDL_POWERSTATE_CHARGED

Plugged in, battery charged

SDL_POWERSTATE_CHARGING

Plugged in, charging battery

SDL_POWERSTATE_ERROR

error determining power status

SDL_POWERSTATE_NO_BATTERY

Plugged in, no battery available

SDL_POWERSTATE_ON_BATTERY

Not plugged in, running on the battery

SDL_POWERSTATE_UNKNOWN

cannot determine power status

Functions

SDL_GetPowerInfo⚠

Get the current power supply details.