sdl3_sys::power
Function SDL_GetPowerInfo

pub unsafe extern "C" fn SDL_GetPowerInfo(
    seconds: *mut c_int,
    percent: *mut c_int,
) -> SDL_PowerState

Get the current power supply details.

You should never take a battery status as absolute truth. Batteries (especially failing batteries) are delicate hardware, and the values reported here are best estimates based on what that hardware reports. It’s not uncommon for older batteries to lose stored power much faster than it reports, or completely drain when reporting it has 20 percent left, etc.

Battery status can change at any time; if you are concerned with power state, you should call this function frequently, and perhaps ignore changes until they seem to be stable for a few seconds.

It’s possible a platform can only report battery percentage or time left but not both.

Parameters

seconds: a pointer filled in with the seconds of battery life left, or NULL to ignore. This will be filled in with -1 if we can’t determine a value or there is no battery.

percent: a pointer filled in with the percentage of battery life left, between 0 and 100, or NULL to ignore. This will be filled in with -1 we can’t determine a value or there is no battery.

Return value

Returns the current battery state or SDL_POWERSTATE_ERROR on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.