sdl3_sys::render
Function SDL_LockTexture

pub unsafe extern "C" fn SDL_LockTexture(
    texture: *mut SDL_Texture,
    rect: *const SDL_Rect,
    pixels: *mut *mut c_void,
    pitch: *mut c_int,
) -> bool

Lock a portion of the texture for write-only pixel access.

As an optimization, the pixels made available for editing don’t necessarily contain the old texture data. This is a write-only operation, and if you need to keep a copy of the texture data you should do that at the application level.

You must use SDL_UnlockTexture() to unlock the pixels and apply any changes.

Parameters

texture: the texture to lock for access, which was created with SDL_TEXTUREACCESS_STREAMING.

rect: an SDL_Rect structure representing the area to lock for access; NULL to lock the entire texture.

pixels: this is filled in with a pointer to the locked pixels, appropriately offset by the locked area.

pitch: this is filled in with the pitch of the locked pixels; the pitch is the length of one row in bytes.

Return value

Returns true on success or false if the texture is not valid or was not created with SDL_TEXTUREACCESS_STREAMING; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_LockTextureToSurface

SDL_UnlockTexture