sdl3_sys::render
Function SDL_SetTextureColorModFloat

pub unsafe extern "C" fn SDL_SetTextureColorModFloat(
    texture: *mut SDL_Texture,
    r: c_float,
    g: c_float,
    b: c_float,
) -> bool

Set an additional color value multiplied into render copy operations.

When this texture is rendered, during the copy operation each source color channel is modulated by the appropriate color value according to the following formula:

srcC = srcC * color

Color modulation is not always supported by the renderer; it will return false if color modulation is not supported.

Parameters

texture: the texture to update.

r: the red color value multiplied into copy operations.

g: the green color value multiplied into copy operations.

b: the blue color value multiplied into copy operations.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetTextureColorModFloat

SDL_SetTextureAlphaModFloat

SDL_SetTextureColorMod