sdl3_sys::scancode
Struct SDL_Scancode

#[repr(transparent)]
pub struct SDL_Scancode(pub c_int);

The SDL keyboard scancode representation.

An SDL scancode is the physical representation of a key on the keyboard, independent of language and keyboard mapping.

Values of this type are used to represent keyboard keys, among other places in the scancode field of the SDL_KeyboardEvent structure.

The values in this enumeration are based on the USB usage page standard: https://usb.org/sites/default/files/hut1_5.pdf

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

UNKNOWN	SDL_SCANCODE_UNKNOWN

A	SDL_SCANCODE_A

B	SDL_SCANCODE_B

C	SDL_SCANCODE_C

D	SDL_SCANCODE_D

E	SDL_SCANCODE_E

F	SDL_SCANCODE_F

G	SDL_SCANCODE_G

H	SDL_SCANCODE_H

I	SDL_SCANCODE_I

J	SDL_SCANCODE_J

K	SDL_SCANCODE_K

L	SDL_SCANCODE_L

M	SDL_SCANCODE_M

N	SDL_SCANCODE_N

O	SDL_SCANCODE_O

P	SDL_SCANCODE_P

Q	SDL_SCANCODE_Q

R	SDL_SCANCODE_R

S	SDL_SCANCODE_S

T	SDL_SCANCODE_T

U	SDL_SCANCODE_U

V	SDL_SCANCODE_V

W	SDL_SCANCODE_W

X	SDL_SCANCODE_X

Y	SDL_SCANCODE_Y

Z	SDL_SCANCODE_Z

_1	SDL_SCANCODE_1

_2	SDL_SCANCODE_2

_3	SDL_SCANCODE_3

_4	SDL_SCANCODE_4

_5	SDL_SCANCODE_5

_6	SDL_SCANCODE_6

_7	SDL_SCANCODE_7

_8	SDL_SCANCODE_8

_9	SDL_SCANCODE_9

_0	SDL_SCANCODE_0

RETURN	SDL_SCANCODE_RETURN

ESCAPE	SDL_SCANCODE_ESCAPE

BACKSPACE	SDL_SCANCODE_BACKSPACE

TAB	SDL_SCANCODE_TAB

SPACE	SDL_SCANCODE_SPACE

MINUS	SDL_SCANCODE_MINUS

EQUALS	SDL_SCANCODE_EQUALS

LEFTBRACKET	SDL_SCANCODE_LEFTBRACKET

RIGHTBRACKET	SDL_SCANCODE_RIGHTBRACKET

BACKSLASH	SDL_SCANCODE_BACKSLASH	Located at the lower left of the return key on ISO keyboards and at the right end of the QWERTY row on ANSI keyboards. Produces REVERSE SOLIDUS (backslash) and VERTICAL LINE in a US layout, REVERSE SOLIDUS and VERTICAL LINE in a UK Mac layout, NUMBER SIGN and TILDE in a UK Windows layout, DOLLAR SIGN and POUND SIGN in a Swiss German layout, NUMBER SIGN and APOSTROPHE in a German layout, GRAVE ACCENT and POUND SIGN in a French Mac layout, and ASTERISK and MICRO SIGN in a French Windows layout.

NONUSHASH	SDL_SCANCODE_NONUSHASH	ISO USB keyboards actually use this code instead of 49 for the same key, but all OSes I’ve seen treat the two codes identically. So, as an implementor, unless your keyboard generates both of those codes and your OS treats them differently, you should generate SDL_SCANCODE_BACKSLASH instead of this code. As a user, you should not rely on this code because SDL will never generate it with most (all?) keyboards.

SEMICOLON	SDL_SCANCODE_SEMICOLON

APOSTROPHE	SDL_SCANCODE_APOSTROPHE

GRAVE	SDL_SCANCODE_GRAVE	Located in the top left corner (on both ANSI and ISO keyboards). Produces GRAVE ACCENT and TILDE in a US Windows layout and in US and UK Mac layouts on ANSI keyboards, GRAVE ACCENT and NOT SIGN in a UK Windows layout, SECTION SIGN and PLUS-MINUS SIGN in US and UK Mac layouts on ISO keyboards, SECTION SIGN and DEGREE SIGN in a Swiss German layout (Mac: only on ISO keyboards), CIRCUMFLEX ACCENT and DEGREE SIGN in a German layout (Mac: only on ISO keyboards), SUPERSCRIPT TWO and TILDE in a French Windows layout, COMMERCIAL AT and NUMBER SIGN in a French Mac layout on ISO keyboards, and LESS-THAN SIGN and GREATER-THAN SIGN in a Swiss German, German, or French Mac layout on ANSI keyboards.

COMMA	SDL_SCANCODE_COMMA

PERIOD	SDL_SCANCODE_PERIOD

SLASH	SDL_SCANCODE_SLASH

CAPSLOCK	SDL_SCANCODE_CAPSLOCK

F1	SDL_SCANCODE_F1

F2	SDL_SCANCODE_F2

F3	SDL_SCANCODE_F3

F4	SDL_SCANCODE_F4

F5	SDL_SCANCODE_F5

F6	SDL_SCANCODE_F6

F7	SDL_SCANCODE_F7

F8	SDL_SCANCODE_F8

F9	SDL_SCANCODE_F9

F10	SDL_SCANCODE_F10

F11	SDL_SCANCODE_F11

F12	SDL_SCANCODE_F12

PRINTSCREEN	SDL_SCANCODE_PRINTSCREEN

SCROLLLOCK	SDL_SCANCODE_SCROLLLOCK

PAUSE	SDL_SCANCODE_PAUSE

INSERT	SDL_SCANCODE_INSERT	insert on PC, help on some Mac keyboards (but does send code 73, not 117)

HOME	SDL_SCANCODE_HOME

PAGEUP	SDL_SCANCODE_PAGEUP

DELETE	SDL_SCANCODE_DELETE

END	SDL_SCANCODE_END

PAGEDOWN	SDL_SCANCODE_PAGEDOWN

RIGHT	SDL_SCANCODE_RIGHT

LEFT	SDL_SCANCODE_LEFT

DOWN	SDL_SCANCODE_DOWN

UP	SDL_SCANCODE_UP

NUMLOCKCLEAR	SDL_SCANCODE_NUMLOCKCLEAR	num lock on PC, clear on Mac keyboards

KP_DIVIDE	SDL_SCANCODE_KP_DIVIDE

KP_MULTIPLY	SDL_SCANCODE_KP_MULTIPLY

KP_MINUS	SDL_SCANCODE_KP_MINUS

KP_PLUS	SDL_SCANCODE_KP_PLUS

KP_ENTER	SDL_SCANCODE_KP_ENTER

KP_1	SDL_SCANCODE_KP_1

KP_2	SDL_SCANCODE_KP_2

KP_3	SDL_SCANCODE_KP_3

KP_4	SDL_SCANCODE_KP_4

KP_5	SDL_SCANCODE_KP_5

KP_6	SDL_SCANCODE_KP_6

KP_7	SDL_SCANCODE_KP_7

KP_8	SDL_SCANCODE_KP_8

KP_9	SDL_SCANCODE_KP_9

KP_0	SDL_SCANCODE_KP_0

KP_PERIOD	SDL_SCANCODE_KP_PERIOD

NONUSBACKSLASH	SDL_SCANCODE_NONUSBACKSLASH	This is the additional key that ISO keyboards have over ANSI ones, located between left shift and Y. Produces GRAVE ACCENT and TILDE in a US or UK Mac layout, REVERSE SOLIDUS (backslash) and VERTICAL LINE in a US or UK Windows layout, and LESS-THAN SIGN and GREATER-THAN SIGN in a Swiss German, German, or French layout.

APPLICATION	SDL_SCANCODE_APPLICATION	windows contextual menu, compose

POWER	SDL_SCANCODE_POWER	The USB document says this is a status flag, not a physical key - but some Mac keyboards do have a power key.

KP_EQUALS	SDL_SCANCODE_KP_EQUALS

F13	SDL_SCANCODE_F13

F14	SDL_SCANCODE_F14

F15	SDL_SCANCODE_F15

F16	SDL_SCANCODE_F16

F17	SDL_SCANCODE_F17

F18	SDL_SCANCODE_F18

F19	SDL_SCANCODE_F19

F20	SDL_SCANCODE_F20

F21	SDL_SCANCODE_F21

F22	SDL_SCANCODE_F22

F23	SDL_SCANCODE_F23

F24	SDL_SCANCODE_F24

EXECUTE	SDL_SCANCODE_EXECUTE

HELP	SDL_SCANCODE_HELP	AL Integrated Help Center

MENU	SDL_SCANCODE_MENU	Menu (show menu)

SELECT	SDL_SCANCODE_SELECT

STOP	SDL_SCANCODE_STOP	AC Stop

AGAIN	SDL_SCANCODE_AGAIN	AC Redo/Repeat

UNDO	SDL_SCANCODE_UNDO	AC Undo

CUT	SDL_SCANCODE_CUT	AC Cut

COPY	SDL_SCANCODE_COPY	AC Copy

PASTE	SDL_SCANCODE_PASTE	AC Paste

FIND	SDL_SCANCODE_FIND	AC Find

MUTE	SDL_SCANCODE_MUTE

VOLUMEUP	SDL_SCANCODE_VOLUMEUP

VOLUMEDOWN	SDL_SCANCODE_VOLUMEDOWN

KP_COMMA	SDL_SCANCODE_KP_COMMA

KP_EQUALSAS400	SDL_SCANCODE_KP_EQUALSAS400

INTERNATIONAL1	SDL_SCANCODE_INTERNATIONAL1	used on Asian keyboards, see footnotes in USB doc

INTERNATIONAL2	SDL_SCANCODE_INTERNATIONAL2

INTERNATIONAL3	SDL_SCANCODE_INTERNATIONAL3	Yen

INTERNATIONAL4	SDL_SCANCODE_INTERNATIONAL4

INTERNATIONAL5	SDL_SCANCODE_INTERNATIONAL5

INTERNATIONAL6	SDL_SCANCODE_INTERNATIONAL6

INTERNATIONAL7	SDL_SCANCODE_INTERNATIONAL7

INTERNATIONAL8	SDL_SCANCODE_INTERNATIONAL8

INTERNATIONAL9	SDL_SCANCODE_INTERNATIONAL9

LANG1	SDL_SCANCODE_LANG1	Hangul/English toggle

LANG2	SDL_SCANCODE_LANG2	Hanja conversion

LANG3	SDL_SCANCODE_LANG3	Katakana

LANG4	SDL_SCANCODE_LANG4	Hiragana

LANG5	SDL_SCANCODE_LANG5	Zenkaku/Hankaku

LANG6	SDL_SCANCODE_LANG6	reserved

LANG7	SDL_SCANCODE_LANG7	reserved

LANG8	SDL_SCANCODE_LANG8	reserved

LANG9	SDL_SCANCODE_LANG9	reserved

ALTERASE	SDL_SCANCODE_ALTERASE	Erase-Eaze

SYSREQ	SDL_SCANCODE_SYSREQ

CANCEL	SDL_SCANCODE_CANCEL	AC Cancel

CLEAR	SDL_SCANCODE_CLEAR

PRIOR	SDL_SCANCODE_PRIOR

RETURN2	SDL_SCANCODE_RETURN2

SEPARATOR	SDL_SCANCODE_SEPARATOR

OUT	SDL_SCANCODE_OUT

OPER	SDL_SCANCODE_OPER

CLEARAGAIN	SDL_SCANCODE_CLEARAGAIN

CRSEL	SDL_SCANCODE_CRSEL

EXSEL	SDL_SCANCODE_EXSEL

KP_00	SDL_SCANCODE_KP_00

KP_000	SDL_SCANCODE_KP_000

THOUSANDSSEPARATOR	SDL_SCANCODE_THOUSANDSSEPARATOR

DECIMALSEPARATOR	SDL_SCANCODE_DECIMALSEPARATOR

CURRENCYUNIT	SDL_SCANCODE_CURRENCYUNIT

CURRENCYSUBUNIT	SDL_SCANCODE_CURRENCYSUBUNIT

KP_LEFTPAREN	SDL_SCANCODE_KP_LEFTPAREN

KP_RIGHTPAREN	SDL_SCANCODE_KP_RIGHTPAREN

KP_LEFTBRACE	SDL_SCANCODE_KP_LEFTBRACE

KP_RIGHTBRACE	SDL_SCANCODE_KP_RIGHTBRACE

KP_TAB	SDL_SCANCODE_KP_TAB

KP_BACKSPACE	SDL_SCANCODE_KP_BACKSPACE

KP_A	SDL_SCANCODE_KP_A

KP_B	SDL_SCANCODE_KP_B

KP_C	SDL_SCANCODE_KP_C

KP_D	SDL_SCANCODE_KP_D

KP_E	SDL_SCANCODE_KP_E

KP_F	SDL_SCANCODE_KP_F

KP_XOR	SDL_SCANCODE_KP_XOR

KP_POWER	SDL_SCANCODE_KP_POWER

KP_PERCENT	SDL_SCANCODE_KP_PERCENT

KP_LESS	SDL_SCANCODE_KP_LESS

KP_GREATER	SDL_SCANCODE_KP_GREATER

KP_AMPERSAND	SDL_SCANCODE_KP_AMPERSAND

KP_DBLAMPERSAND	SDL_SCANCODE_KP_DBLAMPERSAND

KP_VERTICALBAR	SDL_SCANCODE_KP_VERTICALBAR

KP_DBLVERTICALBAR	SDL_SCANCODE_KP_DBLVERTICALBAR

KP_COLON	SDL_SCANCODE_KP_COLON

KP_HASH	SDL_SCANCODE_KP_HASH

KP_SPACE	SDL_SCANCODE_KP_SPACE

KP_AT	SDL_SCANCODE_KP_AT

KP_EXCLAM	SDL_SCANCODE_KP_EXCLAM

KP_MEMSTORE	SDL_SCANCODE_KP_MEMSTORE

KP_MEMRECALL	SDL_SCANCODE_KP_MEMRECALL

KP_MEMCLEAR	SDL_SCANCODE_KP_MEMCLEAR

KP_MEMADD	SDL_SCANCODE_KP_MEMADD

KP_MEMSUBTRACT	SDL_SCANCODE_KP_MEMSUBTRACT

KP_MEMMULTIPLY	SDL_SCANCODE_KP_MEMMULTIPLY

KP_MEMDIVIDE	SDL_SCANCODE_KP_MEMDIVIDE

KP_PLUSMINUS	SDL_SCANCODE_KP_PLUSMINUS

KP_CLEAR	SDL_SCANCODE_KP_CLEAR

KP_CLEARENTRY	SDL_SCANCODE_KP_CLEARENTRY

KP_BINARY	SDL_SCANCODE_KP_BINARY

KP_OCTAL	SDL_SCANCODE_KP_OCTAL

KP_DECIMAL	SDL_SCANCODE_KP_DECIMAL

KP_HEXADECIMAL	SDL_SCANCODE_KP_HEXADECIMAL

LCTRL	SDL_SCANCODE_LCTRL

LSHIFT	SDL_SCANCODE_LSHIFT

LALT	SDL_SCANCODE_LALT	alt, option

LGUI	SDL_SCANCODE_LGUI	windows, command (apple), meta

RCTRL	SDL_SCANCODE_RCTRL

RSHIFT	SDL_SCANCODE_RSHIFT

RALT	SDL_SCANCODE_RALT	alt gr, option

RGUI	SDL_SCANCODE_RGUI	windows, command (apple), meta

MODE	SDL_SCANCODE_MODE	I’m not sure if this is really not covered by any of the above, but since there’s a special SDL_KMOD_MODE for it I’m adding it here

SLEEP	SDL_SCANCODE_SLEEP	Sleep

WAKE	SDL_SCANCODE_WAKE	Wake

CHANNEL_INCREMENT	SDL_SCANCODE_CHANNEL_INCREMENT	Channel Increment

CHANNEL_DECREMENT	SDL_SCANCODE_CHANNEL_DECREMENT	Channel Decrement

MEDIA_PLAY	SDL_SCANCODE_MEDIA_PLAY	Play

MEDIA_PAUSE	SDL_SCANCODE_MEDIA_PAUSE	Pause

MEDIA_RECORD	SDL_SCANCODE_MEDIA_RECORD	Record

MEDIA_FAST_FORWARD	SDL_SCANCODE_MEDIA_FAST_FORWARD	Fast Forward

MEDIA_REWIND	SDL_SCANCODE_MEDIA_REWIND	Rewind

MEDIA_NEXT_TRACK	SDL_SCANCODE_MEDIA_NEXT_TRACK	Next Track

MEDIA_PREVIOUS_TRACK	SDL_SCANCODE_MEDIA_PREVIOUS_TRACK	Previous Track

MEDIA_STOP	SDL_SCANCODE_MEDIA_STOP	Stop

MEDIA_EJECT	SDL_SCANCODE_MEDIA_EJECT	Eject

MEDIA_PLAY_PAUSE	SDL_SCANCODE_MEDIA_PLAY_PAUSE	Play / Pause

MEDIA_SELECT	SDL_SCANCODE_MEDIA_SELECT

AC_NEW	SDL_SCANCODE_AC_NEW	AC New

AC_OPEN	SDL_SCANCODE_AC_OPEN	AC Open

AC_CLOSE	SDL_SCANCODE_AC_CLOSE	AC Close

AC_EXIT	SDL_SCANCODE_AC_EXIT	AC Exit

AC_SAVE	SDL_SCANCODE_AC_SAVE	AC Save

AC_PRINT	SDL_SCANCODE_AC_PRINT	AC Print

AC_PROPERTIES	SDL_SCANCODE_AC_PROPERTIES	AC Properties

AC_SEARCH	SDL_SCANCODE_AC_SEARCH	AC Search

AC_HOME	SDL_SCANCODE_AC_HOME	AC Home

AC_BACK	SDL_SCANCODE_AC_BACK	AC Back

AC_FORWARD	SDL_SCANCODE_AC_FORWARD	AC Forward

AC_STOP	SDL_SCANCODE_AC_STOP	AC Stop

AC_REFRESH	SDL_SCANCODE_AC_REFRESH	AC Refresh

AC_BOOKMARKS	SDL_SCANCODE_AC_BOOKMARKS	AC Bookmarks

SOFTLEFT	SDL_SCANCODE_SOFTLEFT	Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom left of the display.

SOFTRIGHT	SDL_SCANCODE_SOFTRIGHT	Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom right of the display.

CALL	SDL_SCANCODE_CALL	Used for accepting phone calls.

ENDCALL	SDL_SCANCODE_ENDCALL	Used for rejecting phone calls.

RESERVED	SDL_SCANCODE_RESERVED	400-500 reserved for dynamic keycodes

COUNT	SDL_SCANCODE_COUNT	not a key, just marks the number of scancodes for array bounds

Tuple Fields
0: c_int
Implementations

impl SDL_Scancode

pub const UNKNOWN: Self

pub const A: Self

pub const B: Self

pub const C: Self

pub const D: Self

pub const E: Self

pub const F: Self

pub const G: Self

pub const H: Self

pub const I: Self

pub const J: Self

pub const K: Self

pub const L: Self

pub const M: Self

pub const N: Self

pub const O: Self

pub const P: Self

pub const Q: Self

pub const R: Self

pub const S: Self

pub const T: Self

pub const U: Self

pub const V: Self

pub const W: Self

pub const X: Self

pub const Y: Self

pub const Z: Self

pub const _1: Self

pub const _2: Self

pub const _3: Self

pub const _4: Self

pub const _5: Self

pub const _6: Self

pub const _7: Self

pub const _8: Self

pub const _9: Self

pub const _0: Self

pub const RETURN: Self

pub const ESCAPE: Self

pub const BACKSPACE: Self

pub const TAB: Self

pub const SPACE: Self

pub const MINUS: Self

pub const EQUALS: Self

pub const LEFTBRACKET: Self

pub const RIGHTBRACKET: Self

pub const BACKSLASH: Self

Located at the lower left of the return key on ISO keyboards and at the right end of the QWERTY row on ANSI keyboards. Produces REVERSE SOLIDUS (backslash) and VERTICAL LINE in a US layout, REVERSE SOLIDUS and VERTICAL LINE in a UK Mac layout, NUMBER SIGN and TILDE in a UK Windows layout, DOLLAR SIGN and POUND SIGN in a Swiss German layout, NUMBER SIGN and APOSTROPHE in a German layout, GRAVE ACCENT and POUND SIGN in a French Mac layout, and ASTERISK and MICRO SIGN in a French Windows layout.

pub const NONUSHASH: Self

ISO USB keyboards actually use this code instead of 49 for the same key, but all OSes I’ve seen treat the two codes identically. So, as an implementor, unless your keyboard generates both of those codes and your OS treats them differently, you should generate SDL_SCANCODE_BACKSLASH instead of this code. As a user, you should not rely on this code because SDL will never generate it with most (all?) keyboards.

pub const SEMICOLON: Self

pub const APOSTROPHE: Self

pub const GRAVE: Self

Located in the top left corner (on both ANSI and ISO keyboards). Produces GRAVE ACCENT and TILDE in a US Windows layout and in US and UK Mac layouts on ANSI keyboards, GRAVE ACCENT and NOT SIGN in a UK Windows layout, SECTION SIGN and PLUS-MINUS SIGN in US and UK Mac layouts on ISO keyboards, SECTION SIGN and DEGREE SIGN in a Swiss German layout (Mac: only on ISO keyboards), CIRCUMFLEX ACCENT and DEGREE SIGN in a German layout (Mac: only on ISO keyboards), SUPERSCRIPT TWO and TILDE in a French Windows layout, COMMERCIAL AT and NUMBER SIGN in a French Mac layout on ISO keyboards, and LESS-THAN SIGN and GREATER-THAN SIGN in a Swiss German, German, or French Mac layout on ANSI keyboards.

pub const COMMA: Self

pub const PERIOD: Self

pub const SLASH: Self

pub const CAPSLOCK: Self

pub const F1: Self

pub const F2: Self

pub const F3: Self

pub const F4: Self

pub const F5: Self

pub const F6: Self

pub const F7: Self

pub const F8: Self

pub const F9: Self

pub const F10: Self

pub const F11: Self

pub const F12: Self

pub const PRINTSCREEN: Self

pub const SCROLLLOCK: Self

pub const PAUSE: Self

pub const INSERT: Self

insert on PC, help on some Mac keyboards (but does send code 73, not 117)

pub const HOME: Self

pub const PAGEUP: Self

pub const DELETE: Self

pub const END: Self

pub const PAGEDOWN: Self

pub const RIGHT: Self

pub const LEFT: Self

pub const DOWN: Self

pub const UP: Self

pub const NUMLOCKCLEAR: Self

num lock on PC, clear on Mac keyboards

pub const KP_DIVIDE: Self

pub const KP_MULTIPLY: Self

pub const KP_MINUS: Self

pub const KP_PLUS: Self

pub const KP_ENTER: Self

pub const KP_1: Self

pub const KP_2: Self

pub const KP_3: Self

pub const KP_4: Self

pub const KP_5: Self

pub const KP_6: Self

pub const KP_7: Self

pub const KP_8: Self

pub const KP_9: Self

pub const KP_0: Self

pub const KP_PERIOD: Self

pub const NONUSBACKSLASH: Self

This is the additional key that ISO keyboards have over ANSI ones, located between left shift and Y. Produces GRAVE ACCENT and TILDE in a US or UK Mac layout, REVERSE SOLIDUS (backslash) and VERTICAL LINE in a US or UK Windows layout, and LESS-THAN SIGN and GREATER-THAN SIGN in a Swiss German, German, or French layout.

pub const APPLICATION: Self

windows contextual menu, compose

pub const POWER: Self

The USB document says this is a status flag, not a physical key - but some Mac keyboards do have a power key.

pub const KP_EQUALS: Self

pub const F13: Self

pub const F14: Self

pub const F15: Self

pub const F16: Self

pub const F17: Self

pub const F18: Self

pub const F19: Self

pub const F20: Self

pub const F21: Self

pub const F22: Self

pub const F23: Self

pub const F24: Self

pub const EXECUTE: Self

pub const HELP: Self

AL Integrated Help Center

pub const MENU: Self

Menu (show menu)

pub const SELECT: Self

pub const STOP: Self

AC Stop

pub const AGAIN: Self

AC Redo/Repeat

pub const UNDO: Self

AC Undo

pub const CUT: Self

AC Cut

pub const COPY: Self

AC Copy

pub const PASTE: Self

AC Paste

pub const FIND: Self

AC Find

pub const MUTE: Self

pub const VOLUMEUP: Self

pub const VOLUMEDOWN: Self

pub const KP_COMMA: Self

pub const KP_EQUALSAS400: Self

pub const INTERNATIONAL1: Self

used on Asian keyboards, see footnotes in USB doc

pub const INTERNATIONAL2: Self

pub const INTERNATIONAL3: Self

Yen

pub const INTERNATIONAL4: Self

pub const INTERNATIONAL5: Self

pub const INTERNATIONAL6: Self

pub const INTERNATIONAL7: Self

pub const INTERNATIONAL8: Self

pub const INTERNATIONAL9: Self

pub const LANG1: Self

Hangul/English toggle

pub const LANG2: Self

Hanja conversion

pub const LANG3: Self

Katakana

pub const LANG4: Self

Hiragana

pub const LANG5: Self

Zenkaku/Hankaku

pub const LANG6: Self

reserved

pub const LANG7: Self

reserved

pub const LANG8: Self

reserved

pub const LANG9: Self

reserved

pub const ALTERASE: Self

Erase-Eaze

pub const SYSREQ: Self

pub const CANCEL: Self

AC Cancel

pub const CLEAR: Self

pub const PRIOR: Self

pub const RETURN2: Self

pub const SEPARATOR: Self

pub const OUT: Self

pub const OPER: Self

pub const CLEARAGAIN: Self

pub const CRSEL: Self

pub const EXSEL: Self

pub const KP_00: Self

pub const KP_000: Self

pub const THOUSANDSSEPARATOR: Self

pub const DECIMALSEPARATOR: Self

pub const CURRENCYUNIT: Self

pub const CURRENCYSUBUNIT: Self

pub const KP_LEFTPAREN: Self

pub const KP_RIGHTPAREN: Self

pub const KP_LEFTBRACE: Self

pub const KP_RIGHTBRACE: Self

pub const KP_TAB: Self

pub const KP_BACKSPACE: Self

pub const KP_A: Self

pub const KP_B: Self

pub const KP_C: Self

pub const KP_D: Self

pub const KP_E: Self

pub const KP_F: Self

pub const KP_XOR: Self

pub const KP_POWER: Self

pub const KP_PERCENT: Self

pub const KP_LESS: Self

pub const KP_GREATER: Self

pub const KP_AMPERSAND: Self

pub const KP_DBLAMPERSAND: Self

pub const KP_VERTICALBAR: Self

pub const KP_DBLVERTICALBAR: Self

pub const KP_COLON: Self

pub const KP_HASH: Self

pub const KP_SPACE: Self

pub const KP_AT: Self

pub const KP_EXCLAM: Self

pub const KP_MEMSTORE: Self

pub const KP_MEMRECALL: Self

pub const KP_MEMCLEAR: Self

pub const KP_MEMADD: Self

pub const KP_MEMSUBTRACT: Self

pub const KP_MEMMULTIPLY: Self

pub const KP_MEMDIVIDE: Self

pub const KP_PLUSMINUS: Self

pub const KP_CLEAR: Self

pub const KP_CLEARENTRY: Self

pub const KP_BINARY: Self

pub const KP_OCTAL: Self

pub const KP_DECIMAL: Self

pub const KP_HEXADECIMAL: Self

pub const LCTRL: Self

pub const LSHIFT: Self

pub const LALT: Self

alt, option

pub const LGUI: Self

windows, command (apple), meta

pub const RCTRL: Self

pub const RSHIFT: Self

pub const RALT: Self

alt gr, option

pub const RGUI: Self

windows, command (apple), meta

pub const MODE: Self

I’m not sure if this is really not covered by any of the above, but since there’s a special SDL_KMOD_MODE for it I’m adding it here

pub const SLEEP: Self

Sleep

pub const WAKE: Self

Wake

pub const CHANNEL_INCREMENT: Self

Channel Increment

pub const CHANNEL_DECREMENT: Self

Channel Decrement

pub const MEDIA_PLAY: Self

Play

pub const MEDIA_PAUSE: Self

Pause

pub const MEDIA_RECORD: Self

Record

pub const MEDIA_FAST_FORWARD: Self

Fast Forward

pub const MEDIA_REWIND: Self

Rewind

pub const MEDIA_NEXT_TRACK: Self

Next Track

pub const MEDIA_PREVIOUS_TRACK: Self

Previous Track

pub const MEDIA_STOP: Self

Stop

pub const MEDIA_EJECT: Self

Eject

pub const MEDIA_PLAY_PAUSE: Self

Play / Pause

pub const MEDIA_SELECT: Self

pub const AC_NEW: Self

AC New

pub const AC_OPEN: Self

AC Open

pub const AC_CLOSE: Self

AC Close

pub const AC_EXIT: Self

AC Exit

pub const AC_SAVE: Self

AC Save

pub const AC_PRINT: Self

AC Print

pub const AC_PROPERTIES: Self

AC Properties

pub const AC_SEARCH: Self

AC Search

pub const AC_HOME: Self

AC Home

pub const AC_BACK: Self

AC Back

pub const AC_FORWARD: Self

AC Forward

pub const AC_STOP: Self

AC Stop

pub const AC_REFRESH: Self

AC Refresh

pub const AC_BOOKMARKS: Self

AC Bookmarks

pub const SOFTLEFT: Self

Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom left of the display.

pub const SOFTRIGHT: Self

Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom right of the display.

pub const CALL: Self

Used for accepting phone calls.

pub const ENDCALL: Self

Used for rejecting phone calls.

pub const RESERVED: Self

400-500 reserved for dynamic keycodes

pub const COUNT: Self

not a key, just marks the number of scancodes for array bounds