sdl3_sys::stdinc
Function SDL_powf

pub unsafe extern "C" fn SDL_powf(x: c_float, y: c_float) -> c_float

Raise x to the power y

Domain: -INF <= x <= INF, -INF <= y <= INF

Range: -INF <= z <= INF

If y is the base of the natural logarithm (e), consider using SDL_exp instead.

This function operates on single-precision floating point values, use SDL_pow for double-precision floats.

This function may use a different approximation across different versions, platforms and configurations. i.e, it can return a different value given the same input on different machines or operating systems, or if SDL is updated.

Parameters

x: the base.

y: the exponent.

Return value

Returns x raised to the power y.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_pow

SDL_expf

SDL_logf