sdl3_sys::stdinc
Function SDL_rand_r

pub unsafe extern "C" fn SDL_rand_r(
    state: *mut Uint64,
    n: Sint32,
) -> Sint32

Generate a pseudo-random number less than n for positive n

The method used is faster and of better quality than rand() % n. Odds are roughly 99.9% even for n = 1 million. Evenness is better for smaller n, and much worse as n gets bigger.

Example: to simulate a d6 use SDL_rand_r(state, 6) + 1 The +1 converts 0..5 to 1..6

If you want to generate a pseudo-random number in the full range of Sint32, you should use: (Sint32)SDL_rand_bits_r(state)

There are no guarantees as to the quality of the random sequence produced, and this should not be used for security (cryptography, passwords) or where money is on the line (loot-boxes, casinos). There are many random number libraries available with different characteristics and you should pick one of those to meet any serious needs.

Parameters

state: a pointer to the current random number state, this may not be NULL.

n: the number of possible outcomes. n must be positive.

Return value

Returns a random value in the range of [0 .. n-1].

Thread safety

This function is thread-safe, as long as the state pointer isn’t shared between threads.

Availability

This function is available since SDL 3.2.0.

See also

SDL_rand

SDL_rand_bits_r

SDL_randf_r