sdl3_sys::stdinc
Function SDL_uitoa

pub unsafe extern "C" fn SDL_uitoa(
    value: c_uint,
    str: *mut c_char,
    radix: c_int,
) -> *mut c_char

Convert an unsigned integer into a string.

This requires a radix to specified for string format. Specifying 10 produces a decimal number, 16 hexidecimal, etc. Must be in the range of 2 to 36.

Note that this function will overflow a buffer if str is not large enough to hold the output! It may be safer to use SDL_snprintf to clamp output, or SDL_asprintf to allocate a buffer. Otherwise, it doesn’t hurt to allocate much more space than you expect to use (and don’t forget null terminator bytes, etc).

Parameters

value: the unsigned integer to convert.

str: the buffer to write the string into.

radix: the radix to use for string generation.

Return value

Returns str.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_itoa

SDL_ultoa

SDL_ulltoa