sdl3_sys::stdinc
Function SDL_wcslcat

pub unsafe extern "C" fn SDL_wcslcat(
    dst: *mut c_wchar_t,
    src: *const c_wchar_t,
    maxlen: usize,
) -> usize

Concatenate wide strings.

This function appends up to maxlen - SDL_wcslen(dst) - 1 wide characters from src to the end of the wide string in dst, then appends a null terminator.

src and dst must not overlap.

If maxlen - SDL_wcslen(dst) - 1 is less than or equal to 0, then dst is unmodified.

Parameters

dst: The destination buffer already containing the first null-terminated wide string. Must not be NULL and must not overlap with src.

src: The second null-terminated wide string. Must not be NULL, and must not overlap with dst.

maxlen: The length (in wide characters) of the destination buffer.

Return value

Returns the length (in wide characters, excluding the null terminator) of the string in dst plus the length of src.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_wcslcpy