sdl3_sys::stdinc
Function SDL_wcsnlen

pub unsafe extern "C" fn SDL_wcsnlen(
    wstr: *const c_wchar_t,
    maxlen: usize,
) -> usize

This works exactly like wcsnlen() but doesn’t require access to a C runtime.

Counts up to a maximum of maxlen wchar_t values in wstr, excluding the null terminator.

Like SDL_strnlen only counts bytes and not codepoints in a UTF-8 string, this counts wchar_t values in a string, even if the string’s encoding is of variable width, like UTF-16.

Also be aware that wchar_t is different sizes on different platforms (4 bytes on Linux, 2 on Windows, etc).

Also, maxlen is a count of wide characters, not bytes!

Parameters

wstr: The null-terminated wide string to read. Must not be NULL.

maxlen: The maximum amount of wide characters to count.

Return value

Returns the length (in wide characters, excluding the null terminator) of wstr but never more than maxlen.

Thread safety

It is safe to call this function from any thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_wcslen

SDL_utf8strlen

SDL_utf8strnlen