sdl3_sys::surface
Function SDL_MapSurfaceRGB

pub unsafe extern "C" fn SDL_MapSurfaceRGB(
    surface: *mut SDL_Surface,
    r: Uint8,
    g: Uint8,
    b: Uint8,
) -> Uint32

Map an RGB triple to an opaque pixel value for a surface.

This function maps the RGB color value to the specified pixel format and returns the pixel value best approximating the given RGB color value for the given pixel format.

If the surface has a palette, the index of the closest matching color in the palette will be returned.

If the surface pixel format has an alpha component it will be returned as all 1 bits (fully opaque).

If the pixel format bpp (color depth) is less than 32-bpp then the unused upper bits of the return value can safely be ignored (e.g., with a 16-bpp format the return value can be assigned to a Uint16, and similarly a Uint8 for an 8-bpp format).

Parameters

surface: the surface to use for the pixel format and palette.

r: the red component of the pixel in the range 0-255.

g: the green component of the pixel in the range 0-255.

b: the blue component of the pixel in the range 0-255.

Return value

Returns a pixel value.

Availability

This function is available since SDL 3.2.0.

See also

SDL_MapSurfaceRGBA