sdl3_sys::surface
Function SDL_ReadSurfacePixel

pub unsafe extern "C" fn SDL_ReadSurfacePixel(
    surface: *mut SDL_Surface,
    x: c_int,
    y: c_int,
    r: *mut Uint8,
    g: *mut Uint8,
    b: *mut Uint8,
    a: *mut Uint8,
) -> bool

Retrieves a single pixel from a surface.

This function prioritizes correctness over speed: it is suitable for unit tests, but is not intended for use in a game engine.

Like SDL_GetRGBA, this uses the entire 0..255 range when converting color components from pixel formats with less than 8 bits per RGB component.

Parameters

surface: the surface to read.

x: the horizontal coordinate, 0 <= x < width.

y: the vertical coordinate, 0 <= y < height.

r: a pointer filled in with the red channel, 0-255, or NULL to ignore this channel.

g: a pointer filled in with the green channel, 0-255, or NULL to ignore this channel.

b: a pointer filled in with the blue channel, 0-255, or NULL to ignore this channel.

a: a pointer filled in with the alpha channel, 0-255, or NULL to ignore this channel.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.