sdl3_sys
Module system

Platform-specific SDL API functions. These are functions that deal with needs of specific operating systems, that didn’t make sense to offer as platform-independent, generic APIs.

Most apps can make do without these functions, but they can be useful for integrating with other parts of a specific system, adding platform-specific polish to an app, or solving problems that only affect one target.

Structs

SDL_Sandbox

Application sandbox environment.

XTaskQueueObject

XUser

Constants

SDL_ANDROID_EXTERNAL_STORAGE_READ

See the official Android developer guide for more information: http://developer.android.com/guide/topics/data/data-storage.html

SDL_ANDROID_EXTERNAL_STORAGE_WRITE

See the official Android developer guide for more information: http://developer.android.com/guide/topics/data/data-storage.html

SDL_SANDBOX_FLATPAK

SDL_SANDBOX_MACOS

SDL_SANDBOX_NONE

SDL_SANDBOX_SNAP

SDL_SANDBOX_UNKNOWN_CONTAINER

Functions

SDL_GetAndroidActivity⚠

Retrieve the Java instance of the Android activity class.

SDL_GetAndroidCachePath⚠

Get the path used for caching data for this Android application.

SDL_GetAndroidExternalStoragePath⚠

Get the path used for external storage for this Android application.

SDL_GetAndroidExternalStorageState⚠

Get the current state of external storage for this Android application.

SDL_GetAndroidInternalStoragePath⚠

Get the path used for internal storage for this Android application.

SDL_GetAndroidJNIEnv⚠

Get the Android Java Native Interface Environment of the current thread.

SDL_GetAndroidSDKVersion⚠

Query Android API level of the current device.

SDL_GetDXGIOutputInfo⚠

Get the DXGI Adapter and Output indices for the specified display.

SDL_GetDirect3D9AdapterIndex⚠

Get the D3D9 adapter index that matches the specified display.

SDL_GetGDKDefaultUser⚠

Gets a reference to the default user handle for GDK.

SDL_GetGDKTaskQueue⚠

Gets a reference to the global async task queue handle for GDK, initializing if needed.

SDL_GetSandbox⚠

Get the application sandbox environment, if any.

SDL_IsChromebook⚠

Query if the application is running on a Chromebook.

SDL_IsDeXMode⚠

Query if the application is running on a Samsung DeX docking station.

SDL_IsTV⚠

Query if the current device is a TV.

SDL_IsTablet⚠

Query if the current device is a tablet.

SDL_OnApplicationDidChangeStatusBarOrientation⚠

Let iOS apps with external event handling report onApplicationDidChangeStatusBarOrientation.

SDL_OnApplicationDidEnterBackground⚠

Let iOS apps with external event handling report onApplicationDidEnterBackground.

SDL_OnApplicationDidEnterForeground⚠

Let iOS apps with external event handling report onApplicationDidBecomeActive.

SDL_OnApplicationDidReceiveMemoryWarning⚠

Let iOS apps with external event handling report onApplicationDidReceiveMemoryWarning.

SDL_OnApplicationWillEnterBackground⚠

Let iOS apps with external event handling report onApplicationWillResignActive.

SDL_OnApplicationWillEnterForeground⚠

Let iOS apps with external event handling report onApplicationWillEnterForeground.

SDL_OnApplicationWillTerminate⚠

Let iOS apps with external event handling report onApplicationWillTerminate.

SDL_RequestAndroidPermission⚠

Request permissions at runtime, asynchronously.

SDL_SendAndroidBackButton⚠

Trigger the Android system back button behavior.

SDL_SendAndroidMessage⚠

Send a user command to SDLActivity.

SDL_SetLinuxThreadPriority⚠

Sets the UNIX nice value for a thread.

SDL_SetLinuxThreadPriorityAndPolicy⚠

Sets the priority (not nice level) and scheduling policy for a thread.

SDL_SetWindowsMessageHook⚠

Set a callback for every Windows message, run before TranslateMessage().

SDL_SetX11EventHook⚠

Set a callback for every X11 event.

SDL_SetiOSAnimationCallback⚠

Use this function to set the animation callback on Apple iOS.

SDL_SetiOSEventPump⚠

Use this function to enable or disable the SDL event pump on Apple iOS.

SDL_ShowAndroidToast⚠

Shows an Android toast notification.

Type Aliases

MSG

(sdl3-sys) Enable a use-windows-sys-* feature to alias this to MSG from the windows-sys crate. Otherwise it’s an opaque struct.

SDL_RequestAndroidPermissionCallback

Callback that presents a response from a SDL_RequestAndroidPermission call.

SDL_WindowsMessageHook

A callback to be used with SDL_SetWindowsMessageHook.

SDL_X11EventHook

A callback to be used with SDL_SetX11EventHook.

SDL_iOSAnimationCallback

The prototype for an Apple iOS animation callback.

XEvent

(sdl3-sys) Enable either a use-x11-* or a use-x11-dl-* feature to alias this to XEvent from the x11 or x11-dl crates, respectively. Otherwise it’s an opaque struct.

XTaskQueueHandle

XUserHandle