sdl3_sys::video
Function SDL_GetPrimaryDisplay

pub unsafe extern "C" fn SDL_GetPrimaryDisplay() -> SDL_DisplayID

Return the primary display.

Return value

Returns the instance ID of the primary display on success or 0 on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetDisplays