sdl3_sys::video
Function SDL_SetWindowFullscreenMode

pub unsafe extern "C" fn SDL_SetWindowFullscreenMode(
    window: *mut SDL_Window,
    mode: *const SDL_DisplayMode,
) -> bool

Set the display mode to use when a window is visible and fullscreen.

This only affects the display mode used when the window is fullscreen. To change the window size when the window is not fullscreen, use SDL_SetWindowSize().

If the window is currently in the fullscreen state, this request is asynchronous on some windowing systems and the new mode dimensions may not be applied immediately upon the return of this function. If an immediate change is required, call SDL_SyncWindow() to block until the changes have taken effect.

When the new mode takes effect, an SDL_EVENT_WINDOW_RESIZED and/or an SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED event will be emitted with the new mode dimensions.

Parameters

window: the window to affect.

mode: a pointer to the display mode to use, which can be NULL for borderless fullscreen desktop mode, or one of the fullscreen modes returned by SDL_GetFullscreenDisplayModes() to set an exclusive fullscreen mode.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetWindowFullscreenMode

SDL_SetWindowFullscreen

SDL_SyncWindow