sdl3_sys::video
Function SDL_SetWindowKeyboardGrab

pub unsafe extern "C" fn SDL_SetWindowKeyboardGrab(
    window: *mut SDL_Window,
    grabbed: bool,
) -> bool

Set a window’s keyboard grab mode.

Keyboard grab enables capture of system keyboard shortcuts like Alt+Tab or the Meta/Super key. Note that not all system keyboard shortcuts can be captured by applications (one example is Ctrl+Alt+Del on Windows).

This is primarily intended for specialized applications such as VNC clients or VM frontends. Normal games should not use keyboard grab.

When keyboard grab is enabled, SDL will continue to handle Alt+Tab when the window is full-screen to ensure the user is not trapped in your application. If you have a custom keyboard shortcut to exit fullscreen mode, you may suppress this behavior with SDL_HINT_ALLOW_ALT_TAB_WHILE_GRABBED.

If the caller enables a grab while another window is currently grabbed, the other window loses its grab in favor of the caller’s window.

Parameters

window: the window for which the keyboard grab mode should be set.

grabbed: this is true to grab keyboard, and false to release.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetWindowKeyboardGrab

SDL_SetWindowMouseGrab