sdl3_sys::video
Function SDL_SetWindowParent

pub unsafe extern "C" fn SDL_SetWindowParent(
    window: *mut SDL_Window,
    parent: *mut SDL_Window,
) -> bool

Set the window as a child of a parent window.

If the window is already the child of an existing window, it will be reparented to the new owner. Setting the parent window to NULL unparents the window and removes child window status.

If a parent window is hidden or destroyed, the operation will be recursively applied to child windows. Child windows hidden with the parent that did not have their hidden status explicitly set will be restored when the parent is shown.

Attempting to set the parent of a window that is currently in the modal state will fail. Use SDL_SetWindowModal() to cancel the modal status before attempting to change the parent.

Popup windows cannot change parents and attempts to do so will fail.

Setting a parent window that is currently the sibling or descendent of the child window results in undefined behavior.

Parameters

window: the window that should become the child of a parent.

parent: the new parent window for the child window.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_SetWindowModal