sdl3_sys::video
Function SDL_SetWindowShape

pub unsafe extern "C" fn SDL_SetWindowShape(
    window: *mut SDL_Window,
    shape: *mut SDL_Surface,
) -> bool

Set the shape of a transparent window.

This sets the alpha channel of a transparent window and any fully transparent areas are also transparent to mouse clicks. If you are using something besides the SDL render API, then you are responsible for drawing the alpha channel of the window to match the shape alpha channel to get consistent cross-platform results.

The shape is copied inside this function, so you can free it afterwards. If your shape surface changes, you should call SDL_SetWindowShape() again to update the window. This is an expensive operation, so should be done sparingly.

The window must have been created with the SDL_WINDOW_TRANSPARENT flag.

Parameters

window: the window.

shape: the surface representing the shape of the window, or NULL to remove any current shape.

Return value

Returns true on success or false on failure; call SDL_GetError() for more information.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.