sdl3_sys::video
Function SDL_SyncWindow

pub unsafe extern "C" fn SDL_SyncWindow(
    window: *mut SDL_Window,
) -> bool

Block until any pending window state is finalized.

On asynchronous windowing systems, this acts as a synchronization barrier for pending window state. It will attempt to wait until any pending window state has been applied and is guaranteed to return within finite time. Note that for how long it can potentially block depends on the underlying window system, as window state changes may involve somewhat lengthy animations that must complete before the window is in its final requested state.

On windowing systems where changes are immediate, this does nothing.

Parameters

window: the window for which to wait for the pending state to be applied.

Return value

Returns true on success or false if the operation timed out before the window was in the requested state.

Thread safety

This function should only be called on the main thread.

Availability

This function is available since SDL 3.2.0.

See also

SDL_SetWindowSize

SDL_SetWindowPosition

SDL_SetWindowFullscreen

SDL_MinimizeWindow

SDL_MaximizeWindow

SDL_RestoreWindow

SDL_HINT_VIDEO_SYNC_WINDOW_OPERATIONS