sdl3_sys::video
Struct SDL_GLAttr

#[repr(transparent)]
pub struct SDL_GLAttr(pub c_int);

An enumeration of OpenGL configuration attributes.

While you can set most OpenGL attributes normally, the attributes listed above must be known before SDL creates the window that will be used with the OpenGL context. These attributes are set and read with SDL_GL_SetAttribute() and SDL_GL_GetAttribute().

In some cases, these attributes are minimum requests; the GL does not promise to give you exactly what you asked for. It’s possible to ask for a 16-bit depth buffer and get a 24-bit one instead, for example, or to ask for no stencil buffer and still have one available. Context creation should fail if the GL can’t provide your requested attributes at a minimum, but you should check to see exactly what you got.

Availability

This enum is available since SDL 3.2.0.

Known values (sdl3-sys)

Associated constant	Global constant	Description

RED_SIZE	SDL_GL_RED_SIZE	the minimum number of bits for the red channel of the color buffer; defaults to 3.

GREEN_SIZE	SDL_GL_GREEN_SIZE	the minimum number of bits for the green channel of the color buffer; defaults to 3.

BLUE_SIZE	SDL_GL_BLUE_SIZE	the minimum number of bits for the blue channel of the color buffer; defaults to 2.

ALPHA_SIZE	SDL_GL_ALPHA_SIZE	the minimum number of bits for the alpha channel of the color buffer; defaults to 0.

BUFFER_SIZE	SDL_GL_BUFFER_SIZE	the minimum number of bits for frame buffer size; defaults to 0.

DOUBLEBUFFER	SDL_GL_DOUBLEBUFFER	whether the output is single or double buffered; defaults to double buffering on.

DEPTH_SIZE	SDL_GL_DEPTH_SIZE	the minimum number of bits in the depth buffer; defaults to 16.

STENCIL_SIZE	SDL_GL_STENCIL_SIZE	the minimum number of bits in the stencil buffer; defaults to 0.

ACCUM_RED_SIZE	SDL_GL_ACCUM_RED_SIZE	the minimum number of bits for the red channel of the accumulation buffer; defaults to 0.

ACCUM_GREEN_SIZE	SDL_GL_ACCUM_GREEN_SIZE	the minimum number of bits for the green channel of the accumulation buffer; defaults to 0.

ACCUM_BLUE_SIZE	SDL_GL_ACCUM_BLUE_SIZE	the minimum number of bits for the blue channel of the accumulation buffer; defaults to 0.

ACCUM_ALPHA_SIZE	SDL_GL_ACCUM_ALPHA_SIZE	the minimum number of bits for the alpha channel of the accumulation buffer; defaults to 0.

STEREO	SDL_GL_STEREO	whether the output is stereo 3D; defaults to off.

MULTISAMPLEBUFFERS	SDL_GL_MULTISAMPLEBUFFERS	the number of buffers used for multisample anti-aliasing; defaults to 0.

MULTISAMPLESAMPLES	SDL_GL_MULTISAMPLESAMPLES	the number of samples used around the current pixel used for multisample anti-aliasing.

ACCELERATED_VISUAL	SDL_GL_ACCELERATED_VISUAL	set to 1 to require hardware acceleration, set to 0 to force software rendering; defaults to allow either.

RETAINED_BACKING	SDL_GL_RETAINED_BACKING	not used (deprecated).

CONTEXT_MAJOR_VERSION	SDL_GL_CONTEXT_MAJOR_VERSION	OpenGL context major version.

CONTEXT_MINOR_VERSION	SDL_GL_CONTEXT_MINOR_VERSION	OpenGL context minor version.

CONTEXT_FLAGS	SDL_GL_CONTEXT_FLAGS	some combination of 0 or more of elements of the SDL_GLContextFlag enumeration; defaults to 0.

CONTEXT_PROFILE_MASK	SDL_GL_CONTEXT_PROFILE_MASK	type of GL context (Core, Compatibility, ES). See SDL_GLProfile; default value depends on platform.

SHARE_WITH_CURRENT_CONTEXT	SDL_GL_SHARE_WITH_CURRENT_CONTEXT	OpenGL context sharing; defaults to 0.

FRAMEBUFFER_SRGB_CAPABLE	SDL_GL_FRAMEBUFFER_SRGB_CAPABLE	requests sRGB capable visual; defaults to 0.

CONTEXT_RELEASE_BEHAVIOR	SDL_GL_CONTEXT_RELEASE_BEHAVIOR	sets context the release behavior. See SDL_GLContextReleaseFlag; defaults to FLUSH.

CONTEXT_RESET_NOTIFICATION	SDL_GL_CONTEXT_RESET_NOTIFICATION	set context reset notification. See SDL_GLContextResetNotification; defaults to NO_NOTIFICATION.

CONTEXT_NO_ERROR	SDL_GL_CONTEXT_NO_ERROR

FLOATBUFFERS	SDL_GL_FLOATBUFFERS

EGL_PLATFORM	SDL_GL_EGL_PLATFORM

Tuple Fields
0: c_int
Implementations

impl SDL_GLAttr

pub const RED_SIZE: Self

the minimum number of bits for the red channel of the color buffer; defaults to 3.

pub const GREEN_SIZE: Self

the minimum number of bits for the green channel of the color buffer; defaults to 3.

pub const BLUE_SIZE: Self

the minimum number of bits for the blue channel of the color buffer; defaults to 2.

pub const ALPHA_SIZE: Self

the minimum number of bits for the alpha channel of the color buffer; defaults to 0.

pub const BUFFER_SIZE: Self

the minimum number of bits for frame buffer size; defaults to 0.

pub const DOUBLEBUFFER: Self

whether the output is single or double buffered; defaults to double buffering on.

pub const DEPTH_SIZE: Self

the minimum number of bits in the depth buffer; defaults to 16.

pub const STENCIL_SIZE: Self

the minimum number of bits in the stencil buffer; defaults to 0.

pub const ACCUM_RED_SIZE: Self

the minimum number of bits for the red channel of the accumulation buffer; defaults to 0.

pub const ACCUM_GREEN_SIZE: Self

the minimum number of bits for the green channel of the accumulation buffer; defaults to 0.

pub const ACCUM_BLUE_SIZE: Self

the minimum number of bits for the blue channel of the accumulation buffer; defaults to 0.

pub const ACCUM_ALPHA_SIZE: Self

the minimum number of bits for the alpha channel of the accumulation buffer; defaults to 0.

pub const STEREO: Self

whether the output is stereo 3D; defaults to off.

pub const MULTISAMPLEBUFFERS: Self

the number of buffers used for multisample anti-aliasing; defaults to 0.

pub const MULTISAMPLESAMPLES: Self

the number of samples used around the current pixel used for multisample anti-aliasing.

pub const ACCELERATED_VISUAL: Self

set to 1 to require hardware acceleration, set to 0 to force software rendering; defaults to allow either.

pub const RETAINED_BACKING: Self

not used (deprecated).

pub const CONTEXT_MAJOR_VERSION: Self

OpenGL context major version.

pub const CONTEXT_MINOR_VERSION: Self

OpenGL context minor version.

pub const CONTEXT_FLAGS: Self

some combination of 0 or more of elements of the SDL_GLContextFlag enumeration; defaults to 0.

pub const CONTEXT_PROFILE_MASK: Self

type of GL context (Core, Compatibility, ES). See SDL_GLProfile; default value depends on platform.

pub const SHARE_WITH_CURRENT_CONTEXT: Self

OpenGL context sharing; defaults to 0.

pub const FRAMEBUFFER_SRGB_CAPABLE: Self

requests sRGB capable visual; defaults to 0.

pub const CONTEXT_RELEASE_BEHAVIOR: Self

sets context the release behavior. See SDL_GLContextReleaseFlag; defaults to FLUSH.

pub const CONTEXT_RESET_NOTIFICATION: Self

set context reset notification. See SDL_GLContextResetNotification; defaults to NO_NOTIFICATION.

pub const CONTEXT_NO_ERROR: Self

pub const FLOATBUFFERS: Self

pub const EGL_PLATFORM: Self