sdl3_sys::vulkan
Function SDL_Vulkan_GetInstanceExtensions

pub unsafe extern "C" fn SDL_Vulkan_GetInstanceExtensions(
    count: *mut Uint32,
) -> *const *const c_char

Get the Vulkan instance extensions needed for vkCreateInstance.

This should be called after either calling SDL_Vulkan_LoadLibrary() or creating an SDL_Window with the SDL_WINDOW_VULKAN flag.

On return, the variable pointed to by count will be set to the number of elements returned, suitable for using with VkInstanceCreateInfo::enabledExtensionCount, and the returned array can be used with VkInstanceCreateInfo::ppEnabledExtensionNames, for calling Vulkan’s vkCreateInstance API.

You should not free the returned array; it is owned by SDL.

Parameters

count: a pointer filled in with the number of extensions returned.

Return value

Returns an array of extension name strings on success, NULL on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.

See also

SDL_Vulkan_CreateSurface