sdl3_sys::gamepad
Function SDL_GetGamepadStringForType

pub unsafe extern "C" fn SDL_GetGamepadStringForType(
    type: SDL_GamepadType,
) -> *const c_char

Convert from an SDL_GamepadType enum to a string.

Parameters

type: an enum value for a given SDL_GamepadType.

Return value

Returns a string for the given type, or NULL if an invalid type is specified. The string returned is of the format used by SDL_Gamepad mapping strings.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetGamepadTypeFromString