sdl3_sys::mutex
Function SDL_TryLockRWLockForReading

pub unsafe extern "C" fn SDL_TryLockRWLockForReading(
    rwlock: *mut SDL_RWLock,
) -> bool

Try to lock a read/write lock for reading without blocking.

This works just like SDL_LockRWLockForReading(), but if the rwlock is not available, then this function returns false immediately.

This technique is useful if you need access to a resource but don’t want to wait for it, and will return to it to try again later.

Trying to lock for read-only access can succeed if other threads are holding read-only locks, as this won’t prevent access.

This function returns true if passed a NULL rwlock.

Parameters

rwlock: the rwlock to try to lock.

Return value

Returns true on success, false if the lock would block.

Availability

This function is available since SDL 3.2.0.

See also

SDL_LockRWLockForReading

SDL_TryLockRWLockForWriting

SDL_UnlockRWLock