sdl3_sys::pixels
Function SDL_GetRGB

pub unsafe extern "C" fn SDL_GetRGB(
    pixel: Uint32,
    format: *const SDL_PixelFormatDetails,
    palette: *const SDL_Palette,
    r: *mut Uint8,
    g: *mut Uint8,
    b: *mut Uint8,
)

Get RGB values from a pixel in the specified format.

This function uses the entire 8-bit [0..255] range when converting color components from pixel formats with less than 8-bits per RGB component (e.g., a completely white pixel in 16-bit RGB565 format would return [0xff, 0xff, 0xff] not [0xf8, 0xfc, 0xf8]).

Parameters

pixel: a pixel value.

format: a pointer to SDL_PixelFormatDetails describing the pixel format.

palette: an optional palette for indexed formats, may be NULL.

r: a pointer filled in with the red component, may be NULL.

g: a pointer filled in with the green component, may be NULL.

b: a pointer filled in with the blue component, may be NULL.

Thread safety

It is safe to call this function from any thread, as long as the palette is not modified.

Availability

This function is available since SDL 3.2.0.

See also

SDL_GetPixelFormatDetails

SDL_GetRGBA

SDL_MapRGB

SDL_MapRGBA