sdl3_sys::surface
Function SDL_ConvertSurfaceAndColorspace

pub unsafe extern "C" fn SDL_ConvertSurfaceAndColorspace(
    surface: *mut SDL_Surface,
    format: SDL_PixelFormat,
    palette: *mut SDL_Palette,
    colorspace: SDL_Colorspace,
    props: SDL_PropertiesID,
) -> *mut SDL_Surface

Copy an existing surface to a new surface of the specified format and colorspace.

This function converts an existing surface to a new format and colorspace and returns the new surface. This will perform any pixel format and colorspace conversion needed.

If the original surface has alternate images, the new surface will have a reference to them as well.

Parameters

surface: the existing SDL_Surface structure to convert.

format: the new pixel format.

palette: an optional palette to use for indexed formats, may be NULL.

colorspace: the new colorspace.

props: an SDL_PropertiesID with additional color properties, or 0.

Return value

Returns the new SDL_Surface structure that is created or NULL on failure; call SDL_GetError() for more information.

Availability

This function is available since SDL 3.2.0.

See also

SDL_ConvertSurface

SDL_DestroySurface