sdl3_sys::tray
Function SDL_CreateTrayMenu

pub unsafe extern "C" fn SDL_CreateTrayMenu(
    tray: *mut SDL_Tray,
) -> *mut SDL_TrayMenu

Create a menu for a system tray.

This should be called at most once per tray icon.

This function does the same thing as SDL_CreateTraySubmenu(), except that it takes a SDL_Tray instead of a SDL_TrayEntry.

A menu does not need to be destroyed; it will be destroyed with the tray.

Parameters

tray: the tray to bind the menu to.

Return value

Returns the newly created menu.

Thread safety

This function should be called on the thread that created the tray.

Availability

This function is available since SDL 3.2.0.

See also

SDL_CreateTray

SDL_GetTrayMenu

SDL_GetTrayMenuParentTray