sdl3_sys::vulkan
Function SDL_Vulkan_GetPresentationSupport

pub unsafe extern "C" fn SDL_Vulkan_GetPresentationSupport(
    instance: VkInstance,
    physicalDevice: VkPhysicalDevice,
    queueFamilyIndex: Uint32,
) -> bool

Query support for presentation via a given physical device and queue family.

The instance must have been created with extensions returned by SDL_Vulkan_GetInstanceExtensions() enabled.

Parameters

instance: the Vulkan instance handle.

physicalDevice: a valid Vulkan physical device handle.

queueFamilyIndex: a valid queue family index for the given physical device.

Return value

Returns true if supported, false if unsupported or an error occurred.

Availability

This function is available since SDL 3.2.0.

See also

SDL_Vulkan_GetInstanceExtensions